% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{add_external_resources}
\alias{add_external_resources}
\title{Add External Resources}
\usage{
add_external_resources()
}
\description{
Add external resources to the shiny app.

This function performs the following:
\itemize{
\item Adds this package's installed \code{www} folder to the shiny app resource paths
via \code{\link[shiny:resourcePaths]{shiny::addResourcePath()}}.
}
}
