% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_info}
\alias{app_info}
\title{App Information}
\usage{
app_info(...)
}
\arguments{
\item{...}{character vectors, specifying the information to retrieve.}
}
\value{
A list of information about the app.
}
\description{
This function provides information about the app.
}
\examples{
app_info()
}
