% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{app_opts}
\alias{app_opts}
\title{Set Shiny App Options}
\usage{
app_opts(...)
}
\arguments{
\item{...}{Named options to set.}
}
\value{
Returns previously set options invisibly.
}
\description{
Set Shiny App Options.
}
