% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Shiny App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{The input, output, and session functions.}
}
\value{
The server for the Shiny application.
}
\description{
This is the main server function for the Shiny application.
}
\concept{\strong{Shiny App}}
