% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sidebar.R
\name{app_sidebar}
\alias{app_sidebar}
\alias{app_sidebar_ui}
\title{App Sidebar Module}
\usage{
app_sidebar_ui(id)
}
\arguments{
\item{id}{The module's namespace ID.}
}
\value{
\itemize{
\item \code{app_sidebar_ui()}: UI function for the sidebar module.
\item \code{app_sidebar_server()}: Server function for the sidebar module.
}
}
\description{
This module creates the global application sidebar containing filters, options,
user profile, and high-level application metadata information.

It includes the following functions:
\itemize{
\item \code{app_sidebar_ui()}: UI function for the sidebar module.
\item \code{app_sidebar_server()}: Server function for the sidebar module.
}

and the following sub-modules:
\itemize{
\item \code{mod_sidebar_user_profile_ui()}: UI function for the user profile section.
\item \code{mod_sidebar_user_profile_server()}: Server function for the user profile section.
\item \code{mod_sidebar_filters_ui()}: UI function for the filters section.
\item \code{mod_sidebar_filters_server()}: Server function for the filters section.
}
}
\seealso{
\code{\link[=app_ui]{app_ui()}}
}
