% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Shiny App UI}
\usage{
app_ui(req)
}
\arguments{
\item{req}{(internal) The request object.}
}
\value{
The UI for the Shiny application.
}
\description{
This is the main UI function for the Shiny application.
}
