% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{as_entrata_response}
\alias{as_entrata_response}
\title{Handle Entrata API Response}
\usage{
as_entrata_response(resp, endpoint = NULL, method = NULL)
}
\arguments{
\item{resp}{Response object from httr2}

\item{endpoint}{The API endpoint that was called}

\item{method}{The API method that was used}
}
\description{
Handle Entrata API Response
}
\keyword{internal}
