% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{chart_colors}
\alias{chart_colors}
\title{Chart Colors}
\usage{
chart_colors(...)
}
\arguments{
\item{...}{The colors to return.}
}
\value{
A vector of colors.
}
\description{
Function for GMH chart color palette.
}
\examples{
chart_colors("primary", "secondary")
}
