% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_date}
\alias{check_date}
\title{Check Date}
\usage{
check_date(date, arg = rlang::caller_arg(date), call = rlang::caller_env())
}
\arguments{
\item{date}{A date object.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
The provided date object invisibly if it is valid.
}
\description{
This function checks if the provided date is a valid date object.
}
