% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_check.R, R/utils_checks.R
\name{check_db_conn}
\alias{check_db_conn}
\alias{check_db_pool}
\title{Check Database Connection}
\usage{
check_db_conn(conn, arg = rlang::caller_arg(conn), call = rlang::caller_env())

check_db_pool(conn, arg = rlang::caller_arg(conn), call = rlang::caller_env())

check_db_conn(conn, arg = rlang::caller_arg(conn), call = rlang::caller_env())
}
\arguments{
\item{conn}{A database connection object.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
The provided connection object invisibly if it is valid.

The provided connection object invisibly if it is valid.
}
\description{
This function checks if the provided connection is a valid database
connection. The function will throw an error if the connection is not
a valid DBI or pool connection.

This function checks if the provided connection is a valid database
connection. The function will throw an error if the connection is not
a valid DBI or pool connection.
}
