% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{compile_styles}
\alias{compile_styles}
\title{Compile Styles}
\usage{
compile_styles(
  path = pkg_sys("www", "styles"),
  source_map = TRUE,
  cache = FALSE
)
}
\arguments{
\item{path}{The path to the \code{styles} folder. Default is \code{pkg_sys("www", "styles")}.
This folder must contain a \code{scss} folder with a \code{index.scss} file.}
}
\value{
The path to the compiled CSS file.
}
\description{
Compile SASS files to CSS.
}
\examples{
compile_styles()

}
