% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gcs.R
\name{configure_gcs}
\alias{configure_gcs}
\title{Google Cloud Storage Configuration}
\usage{
configure_gcs(default_bucket = NULL)
}
\arguments{
\item{default_bucket}{The default GCS bucket to use.}
}
\description{
This function configures the Google Cloud Storage (GCS) settings for the package.
}
