% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_custom_sign_in.R
\docType{data}
\name{custom_sign_in_ui}
\alias{custom_sign_in_ui}
\title{Custom Sign In UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
custom_sign_in_ui
}
\value{
Sign In UI
}
\description{
Sign In UI from \code{\link[noClocksAuthR:sign_in_ui_default]{noClocksAuthR::sign_in_ui_default()}}.
}
\keyword{datasets}
