% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\alias{db_disconnect}
\alias{db_checkout}
\alias{db_return}
\title{Connect to GMH Database}
\usage{
db_connect(db_config = get_db_config(), user_id = NULL)

db_disconnect(conn)

db_checkout(conn)

db_return(conn)
}
\arguments{
\item{db_config}{Database configuration values as a list. The configuration
values should include the following keys: \code{host}, \code{port}, \code{dbname}, \code{user},
and \code{password} at a minimum. If left \code{NULL}, will retrieve the configuration
via the \code{\link[=get_db_config]{get_db_config()}} function.}

\item{user_id}{An integer representing the user id to use for the database
connection. If not provided, the function will attempt to retrieve the
user id from the shiny session object. If the function is not running
inside a shiny app, the default user id of \code{0} will be used.}

\item{conn}{A database connection object.}
}
\value{
A \code{pool} object representing the connection pool to the GMH
PostgreSQL database.
}
\description{
Functions to connect, disconnect, and manage the database connection to the
GMH PostgreSQL database.
\itemize{
\item \code{db_connect()}: Connect to the GMH PostgreSQL database.
\item \code{db_disconnect()}: Disconnect from the GMH PostgreSQL database.
\item \code{db_checkout()}: Check out a connection from the database connection pool.
\item \code{db_return()}: Return a connection to the database connection pool.
}
}
\details{
The \code{db_connect()} function establishes a connection pool to the GMH PostgreSQL
database hosted in Google Cloud SQL using the \code{pool} package.

Additionally, the function performs the following:
\itemize{
\item Sets up automatic disconnect and cleanup of the connection pool when the R
or shiny session ends.
\item Sets the \code{app.user_id} database session variable
\item Stores the connection pool in the shiny session object
\item Stores the user id in the shiny session object
\item Sets the \code{db.pool} option to the connection pool
}
}
