% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_create_tbl}
\alias{db_create_tbl}
\title{Create Database Table}
\usage{
db_create_tbl(tbl, pool, overwrite = FALSE)
}
\arguments{
\item{tbl}{Character string of the table name.}

\item{pool}{A database connection pool object.}

\item{schema}{Character string of the schema name.}

\item{df}{Data frame to write to the database.}

\item{...}{Additional arguments passed to \code{DBI::dbWriteTable()}.}
}
\value{
Invisible \code{NULL}.
}
\description{
Create a new table in the database.
}
