% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_gmh.R
\name{db_read_gmh_properties}
\alias{db_read_gmh_properties}
\title{GMH Database Functions}
\usage{
db_read_gmh_properties(
  pool,
  property_ids = NULL,
  partner_ids = NULL,
  portfolio_ids = NULL,
  collect = TRUE
)
}
\description{
GMH Database Functions
}
