% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_check.R
\name{db_tbl_exists}
\alias{db_tbl_exists}
\title{Check if Table Exists}
\usage{
db_tbl_exists(conn, tbl_name, in_schema = NULL)
}
\arguments{
\item{conn}{A database connection object.}

\item{tbl_name}{The name of the table to check.}

\item{in_schema}{The schema in which the table resides.}
}
\value{
A logical value indicating if the table exists.
}
\description{
This function checks if a table exists in the database.
}
