% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update.R
\name{db_update}
\alias{db_update}
\alias{db_update_survey_property_summary}
\title{Database Update Functions}
\usage{
db_update_survey_property_summary(
  pool,
  new_values,
  property_id = NULL,
  competitor_id = NULL,
  user_id = NULL
)
}
\arguments{
\item{pool}{A database connection pool object.}

\item{new_values}{A data frame containing the new values to update.}

\item{property_id}{(Optional) The property ID to update.}

\item{competitor_id}{(Optional) The competitor ID to update.}

\item{user_id}{(Optional) The user ID to update.}
}
\value{
The updated data (invisible).
}
\description{
These functions are used to update data in the database.
}
