% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{decrypt_cfg_file}
\alias{decrypt_cfg_file}
\title{Decrypt Configuration File}
\usage{
decrypt_cfg_file(
  path = ".",
  key = "NOCLOCKS_ENCRYPTION_KEY",
  set_env = TRUE,
  cfg_file = pkg_sys("config/config.yml")
)
}
\arguments{
\item{path}{The path to save the decrypted configuration file.}

\item{key}{The name of the environment variable containing the encryption key.}

\item{set_env}{Logical indicating if the \code{R_CONFIG_FILE} environment variable should be set.}

\item{cfg_file}{The path to the configuration file to decrypt.}
}
\value{
The configuration values, silently.
}
\description{
Decrypts the configuration file using the provided encryption key.
}
