% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_config.R
\name{entrata_config}
\alias{entrata_config}
\alias{get_entrata_config}
\alias{validate_entrata_config}
\title{Entrata API Configuration}
\usage{
get_entrata_config(
  key = NULL,
  file = Sys.getenv("R_CONFIG_FILE", "config.yml"),
  config = Sys.getenv("R_CONFIG_ACTIVE", "default")
)

validate_entrata_config(
  cfg,
  arg = rlang::caller_arg(cfg),
  call = rlang::caller_env()
)
}
\arguments{
\item{key}{A character string representing the configuration key to retrieve.
Defaults to \code{NULL} which returns the full \code{entrata} specific configuration
object as a list.}

\item{file}{Configuration file to read from (defaults to
\code{"config.yml"}). If the file isn't found at the location
specified then parent directories are searched for a file
of the same name.}

\item{config}{Name of configuration to read from. Defaults to
the value of the \code{R_CONFIG_ACTIVE} environment variable
("default" if the variable does not exist).}

\item{cfg}{A list containing the Entrata configuration values or a character
string representing the path to the configuration file.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
\itemize{
\item \code{get_entrata_config()}: Returns a list, or vector (if \code{key} is specified),
corresponding to the contents of the \code{entrata} configuration key's values
(i.e. \code{base_url}, \code{username}, and \code{password}).
\item \code{validate_entrata_config()}: Returns the validated configuration invisibly.
}
}
\description{
Entrata API configuration functions for setting, retrieving, and validating the
Entrata API configuration.
\itemize{
\item \code{get_entrata_config()}: This function retrieves the Entrata configuration
from the specified configuration file and returns the configuration as a list.
You can optionally specify a key to retrieve a specific configuration value.
\item \code{validate_entrata_config()}: This function validates the Entrata configuration
to ensure that all required fields are present and that they are of the correct
type. If the configuration is invalid, an error will be thrown in the caller
function's caller environment.
}
}
\details{
The Entrata API configuration functions are used to retrieve and validate
the Entrata API configuration. The configuration is expected to be stored
in a YAML configuration file under the \code{entrata} key. The configuration file
should contain the following fields:
\itemize{
\item \code{base_url} - The base URL for the Entrata API.
\item \code{username} - The username to use for authenticating with the Entrata API.
\item \code{password} - The password to use for authenticating with the Entrata API.
}

The configuration file should be stored in the project directory and should
be named \code{config.yml}. The configuration file should have the following
structure:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{default:
 entrata:
   base_url: "https://gmhcommunities.entrata.com/api/v1/"
   username: "<Entrata API Username>"
   password: "<Entrata API Password>"
}\if{html}{\out{</div>}}

For validation, the following validations are performed on the configuration:
\enumerate{
\item The Entrata configuration must contain the following keys:
\code{username}, \code{password}, and \code{base_url}.
\item All required keys must not be empty or \code{NULL}.
\item The \code{base_url} key must be a valid \code{entrata.com} domain URL.
}

To validate the credentials provided in the configuration use the
\code{\link[=entrata_status]{entrata_status()}} function which pings the API to check the status.
}
\seealso{
\code{\link[config:get]{config::get()}} - the primary workhorse for retrieving configuration
values from configuration files.

Other Entrata: 
\code{\link{entrata_reports}()}
}
\concept{Configuration}
\concept{Entrata}
