% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports_weekly.R
\name{entrata_lease_execution_report}
\alias{entrata_lease_execution_report}
\title{Entrata Weekly Lease Execution Report}
\usage{
entrata_lease_execution_report(
  property_ids = NULL,
  report_date = NULL,
  ...,
  request_id = NULL,
  report_version = NULL,
  max_retries = 15,
  entrata_config = NULL
)
}
\arguments{
\item{property_ids}{Vector of property IDs to include in the report. Defaults to all properties.}

\item{report_date}{Date to use for the report. Defaults to the current date.}

\item{...}{Additional parameters to pass to the report request.}

\item{request_id}{Request ID for the report. Defaults to the current time stamp
as an integer.}

\item{report_version}{Version of the report to request. Defaults to "1.8".}

\item{max_retries}{Maximum number of retries to attempt when performing the request to
the \verb{/queue} endpoint. Defaults to \code{15}.}

\item{entrata_config}{Entrata configuration object. Defaults to the
global Entrata configuration object.}
}
\value{
A tibble containing the weekly lease execution report data.
}
\description{
This function retrieves the weekly lease execution report data from Entrata.

This data is used in conjunction with the data retrieved via \code{\link[=entrata_pre_lease_report]{entrata_pre_lease_report()}}
to append data for the prior seven days (hence weekly) new leases, renewals,
and total by property.
}
\seealso{
\code{\link[=entrata_pre_lease_report]{entrata_pre_lease_report()}}, \code{\link[=entrata_reports]{entrata_reports()}}
}
