% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entrata_methods_tbl}
\alias{entrata_methods_tbl}
\title{Entrata API Methods}
\format{
A tibble with columns \code{endpoint} (character) and \code{method_name} (character).
}
\usage{
entrata_methods_tbl
}
\description{
A table of available Entrata API methods grouped by their endpoints.
}
\keyword{datasets}
