% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports_pre_lease.R
\name{entrata_pre_lease_report}
\alias{entrata_pre_lease_report}
\title{Entrata Pre-Lease Report}
\usage{
entrata_pre_lease_report(
  property_ids = NULL,
  report_date = Sys.Date(),
  summarize_by = c("unit_type", "property", "floorplan_name", "do_not_summarize"),
  group_by = c("unit_type", "floorplan_name", "lease_term", "do_not_group"),
  consider_pre_leased_on = "332",
  charge_code_detail = 1L,
  space_options = "do_not_show",
  additional_units_shown = "available",
  combine_unit_spaces_with_same_lease = 0,
  consolidate_by = "no_consolidation",
  arrange_by_property = 0,
  subtotals = list("summary", "details"),
  yoy = 1,
  lease_occupancy_types = NULL,
  report_version = c("3.2", "3.3"),
  request_id = NULL,
  max_retries = 10,
  entrata_config = NULL
)
}
\arguments{
\item{property_ids}{A vector of property IDs to include in the report. Defaults
to all properties.}

\item{report_date}{The date to use for the report. Defaults to the current date.
This date will be used to derive the target date for the pre-lease season, which
is always \code{9/1} of the current year or the next year, depending on the current
date's month.}

\item{summarize_by}{A character value representing the \code{summarize_by} report filter
parameter for the pre_lease report. Defaults to \code{"property"}. Other options are
\code{"unit_type"}, \code{"floorplan_name"}, or \code{"do_not_summarize"}.}

\item{group_by}{A character value representing the \code{group_by} report filter parameter
for the pre_lease report. Defaults to \code{"do_not_group"}. Other options are \code{"unit_type"},
\code{"floorplan_name"}, or \code{"lease_term"}, or \code{"do_not_group"}.}

\item{consider_pre_leased_on}{A numeric value representing the \code{consider_pre_leased_on}
report filter parameter for the pre_lease report. Defaults to \code{32} which represents
\code{"Lease:Partially Completed"}. Other options are \code{33}, \code{34}, \code{41}, \code{42}, \code{43}, and \code{44}.}

\item{request_id}{A unique identifier for the request. Defaults to the current
timestamp.}

\item{max_retries}{The maximum number of retries to attempt when performing the request to
the \verb{/queue} endpoint. Defaults to \code{10}.}

\item{entrata_config}{The Entrata configuration object. Defaults to the global
Entrata configuration object.}

\item{...}{Named parameters to pass as additional pre_lease report filter parameters.
Must be valid parameters for the pre_lease report.}
}
\value{
A list containing:
\itemize{
\item \code{summary} - A tibble representing the summary table for the pre_lease report.
\item \code{details} - A tibble representing the details table for the pre_lease report.
\item \code{parameters} - A list of the parameters used in the request.
}
}
\description{
Retrieves the summary and details tables for the \code{pre_lease} report from the
Entrata API. The function dynamically constructs the API request and parses
the response to return clean summary and details tables.
}
