% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_properties.R
\name{entrata_properties}
\alias{entrata_properties}
\title{Call the \verb{/properties} Entrata API Endpoint}
\usage{
entrata_properties(
  method_name = c("getProperties", "getFloorPlans", "getWebsites"),
  method_params = list(propertyIds = NULL),
  request_id = NULL,
  entrata_config = get_entrata_config(),
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{method_name}{The Entrata endpoint method name. Defaults to \code{"getProperties"},
but other methods can be used as needed.}

\item{method_params}{Named list of parameters to pass to the method object
of the request body. Defaults to an empty list, but some methods may
have required method parameters that must be provided.}

\item{request_id}{(Optional) A unique identifier for the request. Defaults to \code{NULL}
which will use the current time as an integer for the request ID.}

\item{entrata_config}{(Optional) An Entrata Configuration list. Defaults to
\code{get_entrata_config()}.}

\item{verbose}{(Optional) Logical indicating if verbose output should be printed.
Defaults to \code{FALSE}.}

\item{progress}{(Optional) Logical indicating if a progress indicator should be
displayed. Defaults to \code{FALSE}.}
}
\value{
A parsed response containing property related data depending on which
endpoint method is used.
}
\description{
Calls the \verb{/properties} endpoint from the Entrata API.

Currently the following "methods" (i.e. operations) are supported:
\itemize{
\item \code{getProperties}
\item \code{getFloorPlans}
}
}
