% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_properties.R
\name{entrata_properties_getFloorPlans}
\alias{entrata_properties_getFloorPlans}
\title{Entrata Properties: Get Floor Plans}
\usage{
entrata_properties_getFloorPlans(
  property_id,
  floorplan_ids = NULL,
  use_property_preferences = NULL,
  include_disabled_floorplans = NULL,
  request_id = NULL,
  entrata_config = get_entrata_config()
)
}
\arguments{
\item{property_id}{(Required) The Entrata Property ID to retrieve the floor plans
for. Must be a valid Entrata property ID.}

\item{floorplan_ids}{(Optional) Character or Integer vector of Entrata Floor Plan IDs.
If left \code{NULL} (the default), will return all floor plans for the specified property.
Note that this argument will be merged into a single comma-separated, character
string before being passed to the API request body.}

\item{use_property_preferences}{(Optional) Logical indicating if property preferences
should be used. Defaults to \code{TRUE}. Note that this argument will be coerced to an
integer string before being passed to the API request body (i.e. TRUE -> \code{"1"}).}

\item{include_disabled_floorplans}{(Optional) Logical indicating if disabled floor plans
should be included. Defaults to \code{FALSE}. Note that this argument will be coerced to an
integer string before being passed to the API request body (i.e. FALSE -> \code{"0"}).}

\item{request_id}{(Optional) A unique identifier for the request. If left as \code{NULL},
(the default), will set a request ID via \code{as.integer(Sys.time())}.}

\item{entrata_config}{(Optional) An Entrata Configuration list. Defaults to
\code{get_entrata_config()}. See \code{\link[=get_entrata_config]{get_entrata_config()}} for more information.}
}
\value{
A parsed response containing floor plan data for the specified property.
}
\description{
Calls the \code{getFloorPlans} method from the \verb{/properties} endpoint of the Entrata API.
}
