% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_properties.R
\name{entrata_properties_getProperties}
\alias{entrata_properties_getProperties}
\title{Entrata Properties: Get Properties}
\usage{
entrata_properties_getProperties(
  property_ids = NULL,
  property_lookup_code = NULL,
  show_all_status = TRUE,
  request_id = NULL,
  entrata_config = get_entrata_config()
)
}
\arguments{
\item{property_ids}{(Optional) Character vector of Entrata Property IDs.
If left \code{NULL} (the default), will return all Entrata properties.
Note that this argument will be merged into a single comma-separated, character
string before being passed to the API request body.}

\item{property_lookup_code}{(Optional) Character vector of Entrata Property Lookup Codes.
Not used in this package.}

\item{show_all_status}{(Optional) Logical indicating if all property statuses should be shown.
Defaults to \code{TRUE}. Note that this argument will be coerced to an integer string before
being passed to the API request body (i.e. TRUE -> \code{"1"}).}

\item{request_id}{(Optional) A unique identifier for the request. If left as \code{NULL},
(the default), will set a request ID via \code{as.integer(Sys.time())}.}

\item{entrata_config}{(Optional) An Entrata Configuration list. Defaults to
\code{get_entrata_config()}. See \code{\link[=get_entrata_config]{get_entrata_config()}} for more information.}
}
\value{
A parsed response containing property related datasets.
}
\description{
Calls the \code{getProperties} method from the \verb{/properties} endpoint of the Entrata API.
}
\seealso{
\code{\link[=entrata_properties]{entrata_properties()}}
}
