% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports.R
\name{entrata_reports}
\alias{entrata_reports}
\title{Entrata \verb{/reports} Endpoint}
\usage{
entrata_reports(
  method_name = c("getReportList", "getReportInfo", "getReportData",
    "getDependentFilter"),
  method_params = list(NULL),
  request_id = NULL,
  entrata_config = get_entrata_config(),
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{method_name}{The Entrata endpoint method name. Defaults to \code{"getReportList"},
but other methods can be used as needed.}

\item{method_params}{Named list of parameters to pass to the method object
of the request body. Defaults to an empty list, but some methods may
have required method parameters that must be provided.}

\item{request_id}{(Optional) A unique identifier for the request. Defaults to \code{NULL}
which will use the current time as an integer for the request ID.}

\item{entrata_config}{(Optional) An Entrata Configuration list. Defaults to
\code{get_entrata_config()}.}

\item{verbose}{(Optional) Logical indicating if verbose output should be printed.
Defaults to \code{FALSE}.}

\item{progress}{(Optional) Logical indicating if a progress indicator should be
displayed. Defaults to \code{FALSE}.}
}
\value{
A parsed response containing report related data depending on which
endpoint method is used.
}
\description{
Functions for interacting with the Entrata \verb{/reports} endpoint and its
corresponding methods.
\itemize{
\item \code{entrata_reports()}: Call the Entrata \verb{/reports} endpoint
\item \code{entrata_reports_list()}: List all available reports via the \code{getReportsList} method.
\item \code{entrata_reports_info()}: Get information about reports via the \code{getReportInfo} method.
\item \code{entrata_reports_data()}: Get a \code{queueId} for a report via the \code{getReportData} method.
}

plus,
\itemize{
\item \code{get_latest_report_version()}: Get the latest version of a specified report.
}
}
\details{
The Entrata \verb{/reports} endpoint provides methods for interacting with
the Entrata reporting system and all of the various pre-built reports
specific to the GMH Communities API tenant.

The \verb{/reports} endpoint provides the following methods:
\itemize{
\item \code{getDependentFilter}: Retrieves dependent report filters for a report.
\item \code{getReportData}: Retrieves a \code{queueId} for the corresponding report data.
\item \code{getReportInfo}: Retrieves information about a report.
\item \code{getReportsList}: Retrieves a list of available reports.
}
}
\seealso{
\code{\link[=entrata_request]{entrata_request()}}, \code{\link[=entrata_response]{entrata_response()}}, \code{\link[=entrata_errors]{entrata_errors()}}

Other Entrata: 
\code{\link{entrata_config}}
}
\concept{Entrata}
\concept{Reports}
