% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_reports.R
\name{entrata_reports_list}
\alias{entrata_reports_list}
\title{Entrata Reports List}
\usage{
entrata_reports_list(
  request_id = NULL,
  entrata_config = get_entrata_config(),
  latest_only = TRUE
)
}
\arguments{
\item{request_id}{(Optional) A unique identifier for the request. Defaults to \code{NULL}
which will use the current time as an integer for the request ID.}

\item{entrata_config}{(Optional) An Entrata Configuration list. Defaults to
\code{get_entrata_config()}.}

\item{latest_only}{Logical. If \code{TRUE}, only the latest version of each report
will be returned. Defaults to \code{TRUE}.}
}
\value{
A tibble containing the list of available reports.
}
\description{
List all available Entrata API reports via the \code{getReportsList} method for
the \verb{/reports} endpoint.
}
