% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_endpoint}
\alias{entrata_req_endpoint}
\title{Set Endpoint Path for Entrata Request}
\usage{
entrata_req_endpoint(req, endpoint)
}
\arguments{
\item{req}{An httr2 request object}

\item{endpoint}{The API endpoint to set}
}
\value{
Modified request object with the endpoint URL path appended
}
\description{
Appends an endpoint path to the request URL.
}
