% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_log}
\alias{entrata_req_log}
\title{Log Entrata Request Details}
\usage{
entrata_req_log(req, log_path = NULL)
}
\arguments{
\item{req}{An httr2 request object}

\item{logger}{Optional logger to use}
}
\value{
The original request object
}
\description{
Logs request information according to the configured logger.
}
