% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_req_perform}
\alias{entrata_req_perform}
\title{Perform Entrata API Request}
\usage{
entrata_req_perform(
  req,
  path = NULL,
  verbosity = getOption("entrata.verbosity", NULL),
  mock = getOption("entrata.mock", NULL),
  log = getOption("entrata.log", NULL),
  error_call = rlang::current_env(),
  ...
)
}
\arguments{
\item{req}{An httr2 request object}

\item{verbosity}{Verbosity level of output}

\item{mock}{Mock request response if needed}

\item{error_call}{Call environment for error reporting}

\item{...}{Additional arguments passed to req_perform}

\item{cache}{Logical indicating if response should be cached}

\item{cache_path}{Path to cache the request}

\item{save}{Logical indicating if request should be saved}

\item{save_path}{Path to save request data}

\item{debug}{Logical for debugging request}

\item{logger}{Logger object}
}
\value{
The response object
}
\description{
Executes a prepared Entrata request with custom options.
}
