% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{entrata_resp_body}
\alias{entrata_resp_body}
\title{Entrata Response Body}
\usage{
entrata_resp_body(resp)
}
\arguments{
\item{resp}{\code{httr2::response} object.}
}
\value{
\code{list} with the response body.
}
\description{
Extract the response body from the Entrata API response.
}
