% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{entrata_resp_body_error}
\alias{entrata_resp_body_error}
\title{Entrata Response Error Body}
\usage{
entrata_resp_body_error(resp)
}
\arguments{
\item{resp}{\code{httr2::response} object.}
}
\value{
the response error body.
}
\description{
Extract the error body and message from the response.
}
