% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{entrata_resp_body_json}
\alias{entrata_resp_body_json}
\title{Entrata Response Body JSON}
\usage{
entrata_resp_body_json(resp)
}
\arguments{
\item{resp}{The Entrata API response object. Must be a valid HTTP response.}
}
\value{
A list containing the request ID, status code, and result object.
}
\description{
This function extracts the JSON response body from an Entrata API response.
}
\details{
This function is used to extract the JSON response body from an Entrata API
response object. The function will check the response object for the expected
structure and return the response body as a list containing the request ID,
status code, and result object.
}
