% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{entrata_resp_body_success}
\alias{entrata_resp_body_success}
\title{Entrata Response Success Body}
\usage{
entrata_resp_body_success(resp)
}
\arguments{
\item{resp}{\code{httr2::response} object.}
}
\value{
the response body.
}
\description{
Extract the body from a successful response.
}
