% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{entrata_resp_is_transient}
\alias{entrata_resp_is_transient}
\title{Determine if the Entrata Response is Transient}
\usage{
entrata_resp_is_transient(resp)
}
\arguments{
\item{resp}{\code{httr2::response} object.}
}
\value{
\code{logical} indicating if the response is transient.
}
\description{
Check if the response from Entrata is transient based on the status code
or headers.
}
