% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_parse.R
\name{entrata_resp_parse_arcodes}
\alias{entrata_resp_parse_arcodes}
\title{Parse Entrata A/R Codes Response}
\usage{
entrata_resp_parse_arcodes(resp)
}
\arguments{
\item{resp}{The Entrata API response object.}
}
\value{
A tibble containing the parsed and merged A/R codes data.
}
\description{
This function parses the response data from the Entrata API \verb{/arcodes} endpoint's \code{getArCodes} method.
}
