% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_parse.R
\name{entrata_resp_parse_floorplans}
\alias{entrata_resp_parse_floorplans}
\title{Parse Entrata Floor Plans Response}
\usage{
entrata_resp_parse_floorplans(resp)
}
\arguments{
\item{resp}{The Entrata API response object.}
}
\value{
A tibble containing the parsed floor plan data.
}
\description{
This function parses the response data from the Entrata API \verb{/properties} endpoint's
\code{getFloorPlans} method into various tibbles containing the response's floor plan-level
information and data.
}
