% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_parse.R
\name{entrata_resp_parse_properties}
\alias{entrata_resp_parse_properties}
\alias{entrata_resp_parse_properties_tbl}
\alias{entrata_resp_parse_properties_addresses_tbl}
\alias{entrata_resp_parse_properties_post_months_tbl}
\alias{entrata_resp_parse_properties_hours_tbl}
\alias{entrata_resp_parse_properties_space_options_tbl}
\alias{entrata_resp_parse_properties_lease_terms_tbl}
\alias{entrata_resp_parse_properties_lease_term_windows_tbl}
\title{Parse Entrata Properties Response}
\usage{
entrata_resp_parse_properties(resp)

entrata_resp_parse_properties_tbl(resp_data)

entrata_resp_parse_properties_addresses_tbl(resp_data)

entrata_resp_parse_properties_post_months_tbl(resp_data)

entrata_resp_parse_properties_hours_tbl(resp_data)

entrata_resp_parse_properties_space_options_tbl(resp_data)

entrata_resp_parse_properties_lease_terms_tbl(resp_data)

entrata_resp_parse_properties_lease_term_windows_tbl(resp_data)
}
\arguments{
\item{resp}{The response object from the Entrata API \verb{/properties} endpoint. Must be a valid
\code{\link[httr2:response]{httr2::response()}} object.}

\item{resp_data}{The initial data tibble containing the extracted, flattened JSON data
from the response object. This argument is only used by the second-level parsing functions.}
}
\value{
\code{entrata_resp_parse_properties()} returns a list containing the following tibbles:
\itemize{
\item \code{properties} - Base property data
\item \code{property_addresses} - Property address data
\item \code{property_post_months} - Property post months
\item \code{property_hours} - Property hours
\item \code{property_space_options} - Property space options
\item \code{property_lease_terms} - Property lease terms
\item \code{property_lease_term_windows} - Property lease term windows
}

\code{entrata_resp_parse_properties_tbl()}: A tibble containing the base property data

\code{entrata_resp_parse_properties_addresses_tbl()}: A tibble containing the property address data

\code{entrata_resp_parse_properties_post_months_tbl()}: A tibble containing the property post months

\code{entrata_resp_parse_properties_hours_tbl()}: A tibble containing the property hours

\code{entrata_resp_parse_properties_space_options_tbl()}: A tibble containing the property space options

\code{entrata_resp_parse_properties_lease_terms_tbl()}: A tibble containing the property lease terms

\code{entrata_resp_parse_properties_lease_term_windows_tbl()}: A tibble containing the property lease term windows
}
\description{
These functions parse the response data from the Entrata API \verb{/properties} endpoint's
\code{getProperties} method into various tibbles containing the response's property-level
information and data.

The primary parsing function is \code{entrata_resp_parse_properties()} which
calls the other parsing functions to extract the various property data components.
}
\details{
The \code{entrata_resp_parse_properties()} function is the main parsing function that
calls the other parsing functions to extract the various property data components.

The other parsing functions are as follows:
\itemize{
\item \code{entrata_resp_parse_properties_tbl()} - Parse the base property data
\item \code{entrata_resp_parse_properties_addresses_tbl()} - Parse property address data
\item \code{entrata_resp_parse_properties_post_months_tbl()} - Parse property post months
\item \code{entrata_resp_parse_properties_hours_tbl()} - Parse property hours
\item \code{entrata_resp_parse_properties_space_options_tbl()} - Parse property space options
\item \code{entrata_resp_parse_properties_lease_terms()} - Parse property lease terms
\item \code{entrata_resp_parse_properties_lease_term_windows()} - Parse property lease term windows
}

These functions all use the \code{resp_data} tibble as the initial data source, where
\code{resp_data} is generated from the following operations on the extracted, nested JSON data:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{resp_data <- resp |>
  httr2::resp_body_json() |>
  purrr::pluck("response", "result", "PhysicalProperty", "Property") |>
  jsonlite::toJSON(auto_unbox = TRUE) |>
  jsonlite::fromJSON(flatten = TRUE) |>
  tibble::as_tibble() |>
  janitor::clean_names()
}\if{html}{\out{</div>}}

which flattens the nested JSON data into a tibble ready to be parsed into
the various property data components.
}
\examples{
\dontrun{
# create and perform the API request and get response
req <- entrata_req_properties()
resp <- entrata_req_perform(req)

# parse the response data into tibbles
resp_parsed <- entrata_resp_parse_properties(resp)

# extract the property data tibble
properties_tbl <- resp_parsed$properties

# extract the property address data tibble
property_addresses_tbl <- resp_parsed$property_addresses

# etc.
}
}
