% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{entrata_resp_retry_after}
\alias{entrata_resp_retry_after}
\title{Entrata Response Retry-After}
\usage{
entrata_resp_retry_after(resp)
}
\arguments{
\item{resp}{\code{httr2::response} object.}
}
\value{
\code{numeric} indicating the time to wait before retrying,
or \code{NULL} if not found.
}
\description{
Get the Retry-After time from the response headers.
}
