% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_status.R
\name{entrata_status}
\alias{entrata_status}
\title{Entrata Status}
\usage{
entrata_status(request_id = NULL, entrata_config = get_entrata_config())
}
\arguments{
\item{request_id}{The request ID. If left \code{NULL}, a unique identifier will be
generated via \code{as.integer(Sys.time())}.}

\item{entrata_config}{Entrata API Configuration values as a list. The configuration
values should include the following values: \code{base_url}, \code{username}, and \code{password}.
Defaults to \code{\link[=get_entrata_config]{get_entrata_config()}}.}
}
\value{
A list containing the response from the Entrata API.
}
\description{
Check the status of the Entrata API by calling the \verb{/status} endpoint's \code{getStatus}
method.
}
