% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sql.R
\name{generate_sql_tbl_ddl}
\alias{generate_sql_tbl_ddl}
\title{Generate SQL Table DDL}
\usage{
generate_sql_tbl_ddl(df, table_name, schema = "public")
}
\arguments{
\item{df}{The data frame to generate the DDL for.}

\item{table_name}{The name of the table.}

\item{schema}{The schema to create the table in.}
}
\value{
A SQL DDL statement.
}
\description{
This function generates a SQL table DDL statement based on the input data frame.
}
