% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_maps.R
\name{geocode_address}
\alias{geocode_address}
\title{Geocode an Address}
\usage{
geocode_address(address, gmaps_api_key = get_gmaps_config("api_key"))
}
\arguments{
\item{address}{The address to geocode.}

\item{gmaps_api_key}{The Google Maps API key to use for geocoding.
If not provided, the function will attempt to retrieve the key via
\code{\link[=get_gmaps_api_key]{get_gmaps_api_key()}}.}
}
\value{
A data frame containing the geocoded address information.
}
\description{
Geocode an address using the Google Maps API.
}
\examples{
geocode_address("1600 Amphitheatre Parkway, Mountain View, CA")
}
