% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{get_academic_year}
\alias{get_academic_year}
\title{Academic Year}
\usage{
get_academic_year(as_of_date = Sys.Date())
}
\description{
Get the academic year based on the current date. The academic year typically
starts in August and ends in July of the following year.

This function assumes that if the date is after August 1st, the academic year
is the current year to the next year. If the date is before August 1st, the
academic year is the previous year to the current year.
}
