% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_closest_match}
\alias{get_closest_match}
\title{Get Closest Match}
\usage{
get_closest_match(name, valid_names)
}
\arguments{
\item{name}{The name to match.}

\item{valid_names}{A vector of valid names.}
}
\value{
The closest match to the name from the list of valid names.
}
\description{
Get the closest match to a name from a list of valid names.
}
