% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{get_db_conn}
\alias{get_db_conn}
\title{Get the Database Connection}
\usage{
get_db_conn()
}
\value{
A \code{pool} object representing the connection pool to the GMH PostgreSQL
}
\description{
This function retrieves the database connection pool from the shiny session
object if the function is running inside a shiny app. If the function is
not running inside a shiny app, it will attempt to retrieve the connection
pool from the global options. If the connection pool is not found, the
function will throw an error.
}
