% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_entrata_config}
\alias{get_entrata_config}
\title{Get Entrata API Configuration}
\usage{
get_entrata_config(
  key = NULL,
  file = Sys.getenv("R_CONFIG_FILE", "config.yml"),
  config = Sys.getenv("R_CONFIG_ACTIVE", "default")
)
}
\arguments{
\item{key}{A character string representing the configuration key to retrieve.
Defaults to \code{NULL} which returns the full \code{entrata} specific configuration
object as a list.}

\item{file}{Configuration file to read from (defaults to
\code{"config.yml"}). If the file isn't found at the location
specified then parent directories are searched for a file
of the same name.}

\item{config}{Name of configuration to read from. Defaults to
the value of the \code{R_CONFIG_ACTIVE} environment variable
("default" if the variable does not exist).}
}
\value{
A list, or vector (if \code{key} is specified), corresponding to the
contents of the \code{entrata} configuration key's values (i.e. \code{base_url},
\code{username}, and \code{password}).
}
\description{
This function retrieves the Entrata configuration from the specified configuration
file and returns the configuration as a list. You can optionally specify a key
to retrieve a specific configuration value.
}
\details{
The Entrata configuration is expected to be stored in a YAML configuration file
under the \code{entrata} key. The configuration file should contain the following
fields:
\itemize{
\item \code{base_url} - The base URL for the Entrata API.
\item \code{username} - The username to use for authenticating with the Entrata API.
\item \code{password} - The password to use for authenticating with the Entrata API.
}

For example, in this package's used \code{config.yml} file, there is the following
\code{"entrata"} specific (default) configuration:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{default:
  entrata:
    base_url: "https://gmhcommunities.entrata.com"
    username: "<Entrata API Username>"
    password: "<Entrata API Password>"
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[config:get]{config::get()}} - the primary workhorse for retrieving configuration
values from configuration files.

\code{\link[=validate_entrata_config]{validate_entrata_config()}} - a function that validates the
Entrata configuration to ensure that all required fields are present
and valid.

Other Configuration: 
\code{\link{validate_entrata_config}()}

Other Entrata: 
\code{\link{validate_entrata_config}()},
\code{\link{validate_entrata_request}()}
}
\concept{Configuration}
\concept{Entrata}
