% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{get_fiscal_year}
\alias{get_fiscal_year}
\title{Fiscal Year}
\usage{
get_fiscal_year(as_of_date = Sys.Date())
}
\description{
Get the fiscal year based on the current date. The fiscal year typically
starts in July and ends in June of the following year.

This function assumes that if the date is after July 1st, the fiscal year
is the current year to the next year. If the date is before July 1st, the
fiscal year is the previous year to the current year.
}
