% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{get_weeks_left_to_lease}
\alias{get_weeks_left_to_lease}
\title{Weeks Left to Lease}
\usage{
get_weeks_left_to_lease(as_of_date = Sys.Date())
}
\arguments{
\item{as_of_date}{Date to use as the reference point for deriving other dates from.
Defaults to the current system date.}
}
\value{
The number of weeks left to lease until the start of the next leasing period.
}
\description{
Determine the number of weeks left to lease until the start of the next
leasing period. This is calculated based on the current date and the
start date of the pre-lease season.
}
\examples{
get_weeks_left_to_lease()

}
\seealso{
\link{pre_lease_season}, \link{date_formatting}

Other Date Utilities: 
\code{\link{date_formatting}},
\code{\link{leasing_period}},
\code{\link{leasing_week}},
\code{\link{month_dates}},
\code{\link{pre_lease_season}},
\code{\link{utils_dates}},
\code{\link{weekly_period}}
}
\concept{Date Utilities}
