% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{get_xl_report_file_name}
\alias{get_xl_report_file_name}
\title{Get Excel Report File Name}
\usage{
get_xl_report_file_name(report_name, report_date = Sys.Date())
}
\arguments{
\item{report_name}{The name of the report.}

\item{report_date}{The date of the report. Defaults to the current date.}
}
\value{
A string containing the file name for the Excel report.
}
\description{
This function generates a file name for an Excel report based on the report name and date.

For example, if the report name is \strong{"GMH Report"} and the date is \strong{"2025-01-01"},
the function will return \strong{"2025-01-01-GMH_Report.xlsx"}.
}
