% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{gmh_colors}
\alias{gmh_colors}
\title{GMH Colors}
\usage{
gmh_colors(...)
}
\arguments{
\item{...}{The colors to return.}
}
\value{
A vector of colors.
}
\description{
Function for GMH color palette.
}
\examples{
gmh_colors("primary", "secondary")
}
