% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmh_competitors_tbl}
\alias{gmh_competitors_tbl}
\title{GMH Competitors}
\format{
A tibble with competitor information.
}
\usage{
gmh_competitors_tbl
}
\description{
A dataset of GMH competitors with relevant details.
}
\keyword{datasets}
