% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmh_partners_tbl}
\alias{gmh_partners_tbl}
\title{GMH Partners}
\format{
A tibble with partner information.
}
\usage{
gmh_partners_tbl
}
\description{
A dataset of GMH partners with relevant details.
}
\keyword{datasets}
