% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmh_portfolios_tbl}
\alias{gmh_portfolios_tbl}
\title{GMH Portfolios}
\format{
A tibble with portfolio information.
}
\usage{
gmh_portfolios_tbl
}
\description{
A dataset of GMH portfolios with relevant information.
}
\keyword{datasets}
