% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x, }{y Values to compare.}
}
\description{
This function is a wrapper for the NULL coalescing operator:
If x if \code{NULL}, return y, else return x.
}
\keyword{internal}
