% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{httr2_checks}
\alias{httr2_checks}
\alias{check_request}
\alias{check_response}
\title{\code{httr2} Checks}
\usage{
check_request(req, arg = rlang::caller_arg(req), call = rlang::caller_env())

check_response(resp, arg = rlang::caller_arg(resp), call = rlang::caller_env())
}
\arguments{
\item{req}{The HTTP response object to check.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{resp}{The HTTP response object to check.}

\item{headers}{The HTTP headers object to check.}
}
\value{
Invisibly returns \code{NULL} if the request/response/headers is valid, otherwise
will throw an error.
}
\description{
These functions check the validity of HTTP request and response objects as
well as HTTP headers by checking the R object's class against the classes
defined in the \code{httr2} package.
}
\seealso{
\code{\link[httr2:request]{httr2::request()}}, \code{\link[httr2:response]{httr2::response()}}, \code{\link[httr2:req_headers]{httr2::req_headers()}},
\code{\link[httr2:resp_headers]{httr2::resp_headers()}}
}
