% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icon_text}
\alias{icon_text}
\title{Icon Text}
\usage{
icon_text(icon, text, .function = shiny::icon)
}
\arguments{
\item{icon}{The icon to display. This can be a character string or an icon
object.}

\item{text}{The text to display next to the icon.}

\item{.function}{The function to use to create the icon. Default is \code{shiny::icon}.}
}
\value{
A \code{tagList} object containing the icon and text.
}
\description{
This function is used to create a simple icon and text element.
}
\examples{
icon_text("user", "User")
icon_text(shiny::icon("user"), "User")

}
