% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_integer_string}
\alias{is_integer_string}
\alias{is_integer_string_multi}
\title{Is Integer String}
\usage{
is_integer_string(str, arg = rlang::caller_arg(str))

is_integer_string_multi(str, arg = rlang::caller_arg(str))
}
\arguments{
\item{str}{Character string to check. Typically, with the Entrata API, "multiple"
values are represented as comma separated integer strings.}

\item{arg}{For internal use only, used to capture the name of the argument.}
}
\value{
\code{TRUE} if the string is parseable to an integer, \code{FALSE} otherwise.
}
\description{
Validate a string can be parsed into an integer.

Validate a string with comma separated integers can be parsed into individual integers.
}
\section{Functions}{
\itemize{
\item \code{is_integer_string_multi()}: Is Integer String (Multi)

}}
\examples{
is_integer_string("1")
[1] TRUE

is_integer_string("a")
[1] FALSE
is_integer_string_multi("1,2,3")
[1] TRUE
is_integer_string_multi("1,2,3,")
[1] FALSE
is_integer_string_multi("1,2,a,b")
[1] FALSE
}
