% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{leasing_period}
\alias{leasing_period}
\alias{get_leasing_period}
\alias{get_leasing_period_start_date}
\alias{get_leasing_period_end_date}
\title{Leasing Period}
\usage{
get_leasing_period(as_of_date = Sys.Date())

get_leasing_period_start_date(as_of_date = Sys.Date())

get_leasing_period_end_date(as_of_date = Sys.Date())
}
\arguments{
\item{as_of_date}{Date to use as the reference point for deriving other dates from.
Defaults to the current system date.}
}
\value{
\itemize{
\item \code{get_leasing_period()}: A list containing the start and end dates of the current leasing period.
\item \code{get_leasing_period_start_date()}: The start date of the current leasing period.
\item \code{get_leasing_period_end_date()}: The end date of the current leasing period.
}
}
\description{
These functions handle calculations related to the leasing period for student housing.
The leasing period typically spans an academic year, starting on August 1st
and ending on July 31st of the following year.

The following functions are defined:
\itemize{
\item \code{get_leasing_period()}: Returns the start and end dates of the current leasing period.
\item \code{get_leasing_period_start_date()}: Returns the start date of the current leasing period.
\item \code{get_leasing_period_end_date()}: Returns the end date of the current leasing period.
}
}
\examples{
get_leasing_period()
get_leasing_period_start_date()
get_leasing_period_end_date()

}
\seealso{
\link{pre_lease_season}, \link{date_formatting}

Other Date Utilities: 
\code{\link{date_formatting}},
\code{\link{get_weeks_left_to_lease}()},
\code{\link{leasing_week}},
\code{\link{month_dates}},
\code{\link{pre_lease_season}},
\code{\link{utils_dates}},
\code{\link{weekly_period}}
}
\concept{Date Utilities}
