% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leasing_summary_inputs_tbl}
\alias{leasing_summary_inputs_tbl}
\title{Leasing Summary Inputs}
\format{
A tibble with input field definitions for leasing summaries.
}
\usage{
leasing_summary_inputs_tbl
}
\description{
Input fields for leasing summary data collection in market surveys.
}
\keyword{datasets}
