% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{leasing_week}
\alias{leasing_week}
\alias{get_leasing_week}
\alias{get_leasing_week_start_date}
\alias{get_leasing_week_end_date}
\alias{get_leasing_week_number}
\title{Leasing Week}
\usage{
get_leasing_week(as_of_date = Sys.Date(), week_start = "Monday")

get_leasing_week_start_date(as_of_date = Sys.Date(), week_start = "Monday")

get_leasing_week_end_date(as_of_date = Sys.Date(), week_start = "Monday")

get_leasing_week_number(as_of_date = Sys.Date())
}
\arguments{
\item{as_of_date}{Date to use as the reference point for deriving other dates from.
Defaults to the current system date.}

\item{week_start}{week start day (Default is 7, Sunday. Set \code{lubridate.week.start} to
override). Full or abbreviated names of the days of the week can be in English or
as provided by the current locale.}
}
\value{
\itemize{
\item \code{get_leasing_week()}: A list containing the start and end dates of the leasing week.
\item \code{get_leasing_week_start_date()}: The start date of the leasing week.
\item \code{get_leasing_week_end_date()}: The end date of the leasing week.
\item \code{get_leasing_week_number()}: The week number of the leasing week.
}
}
\description{
Calculates the start and end dates of a leasing week for a given "as of"
date. The leasing week is defined as starting on a specific day of the week
(e.g., Monday, Sunday, etc.). Lubridate supports multiple ways of specifying
\code{week_start}.

The following functions are defined:
\itemize{
\item \code{get_leasing_week()}: Returns both the start and end dates of the leasing week.
\item \code{get_leasing_week_start_date()}: Returns the start date of the leasing week.
\item \code{get_leasing_week_end_date()}: Returns the end date of the leasing week.
\item \code{get_leasing_week_number()}: Returns the week number of the leasing week.
}
}
\examples{
get_leasing_week()
get_leasing_week_start_date("2025-01-15", "Monday")
get_leasing_week_end_date(Sys.Date(), "Sunday")
}
\seealso{
Other Date Utilities: 
\code{\link{date_formatting}},
\code{\link{get_weeks_left_to_lease}()},
\code{\link{leasing_period}},
\code{\link{month_dates}},
\code{\link{pre_lease_season}},
\code{\link{utils_dates}},
\code{\link{weekly_period}}
}
\concept{Date Utilities}
