% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box_score.R
\name{mod_box_score}
\alias{mod_box_score}
\alias{mod_box_score_ui}
\alias{mod_box_score_server}
\alias{mod_box_score_demo}
\title{Box Score Shiny Module}
\usage{
mod_box_score_ui(id)

mod_box_score_server(id, pool = NULL, global_filters = NULL)

mod_box_score_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_box_score_ui()}: UI output
\item \code{mod_box_score_server()}: List of reactive expressions.
\item \code{mod_box_score_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Box Score page.

Includes the following functions:
\itemize{
\item \code{mod_box_score_ui()}: User Interface (UI) for the module.
\item \code{mod_box_score_server()}: Server logic for the module.
\item \code{mod_box_score_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_box_score_demo()
\dontshow{\}) # examplesIf}
}
