% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_excel_report.R
\name{mod_excel_report}
\alias{mod_excel_report}
\alias{mod_excel_report_ui}
\alias{mod_excel_report_server}
\alias{mod_excel_report_demo}
\title{Excel Report Shiny Module}
\usage{
mod_excel_report_ui(id)

mod_excel_report_server(
  id,
  pool = NULL,
  summary_table_data = NULL,
  details_table_data = NULL,
  report_parameters = NULL
)

mod_excel_report_demo()
}
\arguments{
\item{id}{Module's namespace ID.}
}
\value{
\itemize{
\item \code{mod_excel_report_ui()}: UI HTML Output.
\item \code{mod_excel_report_server()}: Reactive values returned from server logic.
}
}
\description{
A Shiny Module for generating and exporting/downloading a custom Excel
Report from a Shiny Application.
\itemize{
\item \code{mod_excel_report_ui()}: User interface including the download button
and any necessary export options.
\item \code{mod_excel_report_server()}: Server logic for generating the Excel
report's \code{\link[shiny:downloadHandler]{shiny::downloadHandler()}} logic.
}
}
\examples{
if (interactive()) {
  mod_excel_report_demo()
}
}
