% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_floorplans.R
\name{mod_floorplans}
\alias{mod_floorplans}
\alias{mod_floorplans_ui}
\alias{mod_floorplans_server}
\alias{mod_floorplans_demo}
\title{Floorplans Shiny Module}
\usage{
mod_floorplans_ui(id)

mod_floorplans_server(id, pool = NULL, global_filters = NULL)

mod_floorplans_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_floorplans_ui()}: UI output
\item \code{mod_floorplans_server()}: List of reactive expressions.
\item \code{mod_floorplans_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Floorplans page.

Includes the following functions:
\itemize{
\item \code{mod_floorplans_ui()}: User Interface (UI) for the module.
\item \code{mod_floorplans_server()}: Server logic for the module.
\item \code{mod_floorplans_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_floorplans_demo()
\dontshow{\}) # examplesIf}
}
