% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{mod_home}
\alias{mod_home}
\alias{mod_home_ui}
\alias{mod_home_server}
\title{Home (Page) Shiny Module}
\usage{
mod_home_ui(id, ...)

mod_home_server(id, filters, data, user)
}
\arguments{
\item{id}{The module id.}
}
\value{
\itemize{
\item \code{mod_home_ui()}: The UI HTML wrapped in a \code{\link[htmltools:tagList]{htmltools::tagList()}}.
\item \code{mod_home_server()}: A shiny server function.
}
}
\description{
A Shiny module for creating the home page tab of the GMH Data Hub's Leasing
Dashboard.

This module includes the UI and Server functions:
\itemize{
\item \code{mod_home_ui()}
\item \code{mod_home_server()}
}
}
