% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_performance.R
\name{mod_performance}
\alias{mod_performance}
\alias{mod_performance_ui}
\alias{mod_performance_server}
\alias{mod_performance_demo}
\title{Performance Shiny Module}
\usage{
mod_performance_ui(id)

mod_performance_server(id, pool = NULL, global_filters = NULL)

mod_performance_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_performance_ui()}: UI output
\item \code{mod_performance_server()}: List of reactive expressions.
\item \code{mod_performance_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Performance page.

Includes the following functions:
\itemize{
\item \code{mod_performance_ui()}: User Interface (UI) for the module.
\item \code{mod_performance_server()}: Server logic for the module.
\item \code{mod_performance_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_performance_demo()
\dontshow{\}) # examplesIf}
}
