% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pre_lease.R
\name{mod_pre_lease}
\alias{mod_pre_lease}
\alias{mod_pre_lease_ui}
\alias{mod_pre_lease_server}
\alias{mod_pre_lease_demo}
\title{Pre Lease Shiny Module}
\usage{
mod_pre_lease_ui(id)

mod_pre_lease_server(id, pool = NULL, global_filters = NULL)

mod_pre_lease_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_pre_lease_ui()}: UI output
\item \code{mod_pre_lease_server()}: List of reactive expressions.
\item \code{mod_pre_lease_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Pre Lease page.

Includes the following functions:
\itemize{
\item \code{mod_pre_lease_ui()}: User Interface (UI) for the module.
\item \code{mod_pre_lease_server()}: Server logic for the module.
\item \code{mod_pre_lease_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_pre_lease_demo()
\dontshow{\}) # examplesIf}
}
