% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_properties.R
\name{mod_properties}
\alias{mod_properties}
\alias{mod_properties_ui}
\title{Properties Shiny Module}
\usage{
mod_properties_ui(id, ...)
}
\arguments{
\item{id}{The module id.}

\item{...}{Additional parameters.}
}
\value{
\itemize{
\item \code{mod_properties_ui()}: Shiny UI.
\item \code{mod_properties_server()}: Shiny Server.
}
}
\description{
This is the shiny module for the properties page/tab in the shiny app.

It includes the UI and server functions:
\itemize{
\item \code{mod_properties_ui()}
\item \code{mod_properties_server()}
}

The module includes the following children modules for its major functional
areas:
\itemize{
\item \link{mod_property_tbl} - For listing/displaying all properties.
\item \link{mod_property_map} - For displaying a map of all properties.
\item \link{mod_property_details} - For displaying the details of a single property.
}
}
