% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_property_units.R
\name{mod_property_units}
\alias{mod_property_units}
\alias{mod_property_units_ui}
\alias{mod_property_units_server}
\alias{mod_property_units_demo}
\title{Property Units Shiny Module}
\usage{
mod_property_units_ui(id)

mod_property_units_server(id, pool = NULL, global_filters = NULL)

mod_property_units_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_property_units_ui()}: UI output
\item \code{mod_property_units_server()}: List of reactive expressions.
\item \code{mod_property_units_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Property Units page.

Includes the following functions:
\itemize{
\item \code{mod_property_units_ui()}: User Interface (UI) for the module.
\item \code{mod_property_units_server()}: Server logic for the module.
\item \code{mod_property_units_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_property_units_demo()
\dontshow{\}) # examplesIf}
}
