% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_residents.R
\name{mod_residents}
\alias{mod_residents}
\alias{mod_residents_ui}
\alias{mod_residents_server}
\alias{mod_residents_demo}
\title{Residents Shiny Module}
\usage{
mod_residents_ui(id)

mod_residents_server(id, pool = NULL, global_filters = NULL)

mod_residents_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_residents_ui()}: UI output
\item \code{mod_residents_server()}: List of reactive expressions.
\item \code{mod_residents_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Residents page.

Includes the following functions:
\itemize{
\item \code{mod_residents_ui()}: User Interface (UI) for the module.
\item \code{mod_residents_server()}: Server logic for the module.
\item \code{mod_residents_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_residents_demo()
\dontshow{\}) # examplesIf}
}
