% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{mod_sidebar}
\alias{mod_sidebar}
\alias{mod_sidebar_ui}
\alias{mod_sidebar_server}
\title{App Sidebar Module}
\usage{
mod_sidebar_ui(id, ...)

mod_sidebar_server(id, ...)
}
\arguments{
\item{id}{The module id.}

\item{...}{Additional parameters.}
}
\value{
\itemize{
\item \code{mod_sidebar_ui()}: A shiny UI function.
\item \code{mod_sidebar_server()}: A shiny server function.
}
}
\description{
This module creates a sidebar for the app that contains filters, options, and about information.

It includes the UI and server functions:
\itemize{
\item \code{mod_sidebar_ui()}
\item \code{mod_sidebar_server()}
}
}
