% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_fees.R
\name{mod_survey_fees}
\alias{mod_survey_fees}
\alias{mod_survey_fees_ui}
\alias{mod_survey_fees_server}
\alias{mod_survey_fees_demo}
\title{Survey Fees Shiny Module}
\usage{
mod_survey_fees_ui(id)

mod_survey_fees_server(
  id,
  pool = NULL,
  survey_data = NULL,
  selected_filters = NULL,
  db_trigger_func = NULL,
  edit_survey_section = NULL
)

mod_survey_fees_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}

\item{survey_data}{Reactive data for the survey.}

\item{selected_filters}{Reactive data for the selected filters.}

\item{db_trigger_func}{Reactive Function to trigger database updates.}

\item{edit_survey_section}{Reactive Function representing the action button to edit the survey section.}

\item{map_data}{Reactive data for the map.}
}
\value{
\itemize{
\item \code{mod_survey_fees_ui()}: UI output
\item \code{mod_survey_fees_server()}: List of reactive expressions.
\item \code{mod_survey_fees_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Survey Fees page.

Includes the following functions:
\itemize{
\item \code{mod_survey_fees_ui()}: User Interface (UI) for the module.
\item \code{mod_survey_fees_server()}: Server logic for the module.
\item \code{mod_survey_fees_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_survey_fees_demo()
\dontshow{\}) # examplesIf}
}
