% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights.R
\name{mod_survey_insights}
\alias{mod_survey_insights}
\alias{mod_survey_insights_ui}
\alias{mod_survey_insights_server}
\alias{mod_survey_insights_demo}
\title{Survey Insights Shiny Module}
\usage{
mod_survey_insights_ui(id)

mod_survey_insights_server(id, pool = NULL)

mod_survey_insights_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_survey_insights_ui()}: UI output
\item \code{mod_survey_insights_server()}: List of reactive expressions.
\item \code{mod_survey_insights_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Survey Insights page.

Includes the following functions:
\itemize{
\item \code{mod_survey_insights_ui()}: User Interface (UI) for the module.
\item \code{mod_survey_insights_server()}: Server logic for the module.
\item \code{mod_survey_insights_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_survey_insights_demo()
\dontshow{\}) # examplesIf}
}
