% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights_comparison.R
\name{mod_survey_insights_comparison}
\alias{mod_survey_insights_comparison}
\alias{mod_survey_insights_comparison_ui}
\alias{mod_survey_insights_comparison_server}
\alias{mod_survey_insights_comparison_demo}
\title{Survey Insights Comparison Module}
\usage{
mod_survey_insights_comparison_ui(id)

mod_survey_insights_comparison_server(id, pool = NULL)

mod_survey_insights_comparison_demo()
}
\arguments{
\item{id}{The module ID.}
}
\value{
\itemize{
\item \code{mod_survey_insights_comparison_ui()}: User interface for the module
\item \code{mod_survey_insights_comparison_server()}: Server logic for the module
\item \code{mod_survey_insights_comparison_demo()}: Demo the module
}
}
\description{
This module provides comparison features related to the market survey data.

The following functions are implemented:
\itemize{
\item \code{mod_survey_insights_comparison_ui()}: User interface for the SWOT Analysis Module
\item \code{mod_survey_insights_comparison_server()}: Server logic for the SWOT Analysis Module
\item \code{mod_survey_insights_comparison_demo()}: Demo the SWOT Analysis Module
}
}
\seealso{
\code{\link[=mod_survey_insights]{mod_survey_insights()}}
}
