% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights_swot.R
\name{mod_survey_insights_swot}
\alias{mod_survey_insights_swot}
\alias{mod_survey_insights_swot_server}
\alias{mod_survey_insights_swot_demo}
\title{SWOT Analysis Module}
\usage{
mod_survey_insights_swot(id)

mod_survey_insights_swot_server(id, pool = NULL)

mod_survey_insights_swot_demo()
}
\arguments{
\item{id}{The module ID.}
}
\value{
\itemize{
\item \code{mod_survey_insights_swot_ui()}: User interface for the SWOT Analysis Module
\item \code{mod_survey_insights_swot_server()}: Server logic for the SWOT Analysis Module
\item \code{mod_survey_insights_swot_demo()}: Demo the SWOT Analysis Module
}
}
\description{
This module provides a mechanism for conducting a SWOT analysis in a Shiny
App and includes functions for the UI and server:
\itemize{
\item \code{mod_survey_insights_swot_ui()}: User interface for the SWOT Analysis Module
\item \code{mod_survey_insights_swot_server()}: Server logic for the SWOT Analysis Module
\item \code{mod_survey_insights_swot_demo()}: Demo the SWOT Analysis Module
}
}
\details{
The SWOT Analysis Module provides a structured approach to evaluating the
internal strengths and weaknesses, and external opportunities and threats
of a business or project. This module is designed to facilitate the
collection and organization of SWOT analysis data for use in strategic
planning and decision-making processes.
}
\seealso{
\code{\link[=mod_survey_insights]{mod_survey_insights()}}
}
