% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_insights_trends.R
\name{mod_survey_insights_trends}
\alias{mod_survey_insights_trends}
\alias{mod_survey_insights_trends_ui}
\alias{mod_survey_insights_trends_server}
\title{Market Survey Trends Module}
\usage{
mod_survey_insights_trends_ui(id)

mod_survey_insights_trends_server(id, pool = NULL)
}
\arguments{
\item{id}{The module ID.}
}
\value{
\itemize{
\item \code{mod_survey_insights_trends_ui()}: UI function for the trends insights module.
\item \code{mod_survey_insights_trends_server()}: Server function for the trends insights module.
\item \code{mod_survey_insights_trends_demo()}: Run a demo app of the module.
}
}
\description{
This module provides insights into market survey trends, including occupancy rates
and rent comparisons.

It includes the following functions:
\itemize{
\item \code{mod_survey_insights_trends_ui()}: UI function for the trends insights module.
\item \code{mod_survey_insights_trends_server()}: Server function for the trends insights module.
\item \code{mod_survey_insights_trends_demo()}: Run a demo app of the module.
}
}
\seealso{
\code{\link[=mod_survey_insights]{mod_survey_insights()}}
}
