% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_parking.R
\name{mod_survey_parking}
\alias{mod_survey_parking}
\alias{mod_survey_parking_ui}
\alias{mod_survey_parking_server}
\alias{mod_survey_parking_demo}
\title{Survey Parking Shiny Module}
\usage{
mod_survey_parking_ui(id)

mod_survey_parking_server(
  id,
  pool = NULL,
  selected_property_id = NULL,
  selected_competitor_id = NULL,
  edit_survey_section = NULL
)

mod_survey_parking_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_survey_parking_ui()}: UI output
\item \code{mod_survey_parking_server()}: List of reactive expressions.
\item \code{mod_survey_parking_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Survey Parking page.

Includes the following functions:
\itemize{
\item \code{mod_survey_parking_ui()}: User Interface (UI) for the module.
\item \code{mod_survey_parking_server()}: Server logic for the module.
\item \code{mod_survey_parking_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_survey_parking_demo()
\dontshow{\}) # examplesIf}
}
