% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_short_term_leases.R
\name{mod_survey_short_term_leases}
\alias{mod_survey_short_term_leases}
\alias{mod_survey_short_term_leases_ui}
\alias{mod_survey_short_term_leases_server}
\alias{mod_survey_short_term_leases_demo}
\title{Survey Short Term Leases Shiny Module}
\usage{
mod_survey_short_term_leases_ui(id)

mod_survey_short_term_leases_server(id, pool = NULL, global_filters = NULL)

mod_survey_short_term_leases_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_survey_short_term_leases_ui()}: UI output
\item \code{mod_survey_short_term_leases_server()}: List of reactive expressions.
\item \code{mod_survey_short_term_leases_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Survey Short Term Leases page.

Includes the following functions:
\itemize{
\item \code{mod_survey_short_term_leases_ui()}: User Interface (UI) for the module.
\item \code{mod_survey_short_term_leases_server()}: Server logic for the module.
\item \code{mod_survey_short_term_leases_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_survey_short_term_leases_demo()
\dontshow{\}) # examplesIf}
}
