% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{month_dates}
\alias{month_dates}
\alias{get_end_of_month}
\alias{get_start_of_month}
\title{Get the Start and End Dates for a Month}
\usage{
get_end_of_month(as_of_date)

get_start_of_month(as_of_date)
}
\arguments{
\item{as_of_date}{Date to use as the reference point for deriving other dates from.
Defaults to the current system date.}
}
\value{
\itemize{
\item \code{get_end_of_month()}: The date of the last day of the month.
\item \code{get_start_of_month()}: The date of the first day of the month.
}
}
\description{
These functions calculate the start and end dates for a given month.

The following functions are defined:
\itemize{
\item \code{get_end_of_month()}: Returns the date of the last day of the month.
\item \code{get_start_of_month()}: Returns the date of the first day of the month.
}
}
\examples{
get_end_of_month()
get_start_of_month()
}
\seealso{
Other Date Utilities: 
\code{\link{date_formatting}},
\code{\link{get_weeks_left_to_lease}()},
\code{\link{leasing_period}},
\code{\link{leasing_week}},
\code{\link{pre_lease_season}},
\code{\link{utils_dates}},
\code{\link{weekly_period}}
}
\concept{Date Utilities}
