% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{pre_lease_season}
\alias{pre_lease_season}
\alias{get_pre_lease_year}
\alias{get_pre_lease_season}
\alias{get_pre_lease_season_start_date}
\alias{get_entrata_custom_pre_lease_date}
\title{Pre-Lease Season and Leasing Cycle}
\usage{
get_pre_lease_year(as_of_date = Sys.Date())

get_pre_lease_season(as_of_date = Sys.Date())

get_pre_lease_season_start_date(as_of_date = Sys.Date())

get_entrata_custom_pre_lease_date(as_of_date = Sys.Date())
}
\arguments{
\item{as_of_date}{Date to use as the reference point for deriving other dates from.
Defaults to the current system date.}
}
\value{
\itemize{
\item \code{get_pre_lease_year()}: The year when the upcoming pre-lease season starts.
\item \code{get_pre_lease_season()}: A list containing the start and end dates of the upcoming pre-lease season.
\item \code{get_pre_lease_season_start_date()}: The start date of the upcoming pre-lease season.
\item \code{get_leasing_period_end_date()}: The end date of the current leasing period.
}
}
\description{
These functions handle calculations related to the pre-lease season and the overall leasing cycle
for student housing. The pre-lease season is the period during which leasing for the upcoming
academic year typically occurs.

The following functions are defined:
\itemize{
\item \code{get_pre_lease_year()}: Returns the year of the upcoming pre-lease season.
\item \code{get_pre_lease_season()}: Returns the start and end dates of the upcoming pre-lease season.
\item \code{get_pre_lease_season_start_date()}: Returns the start date of the upcoming pre-lease season.
\item \code{get_leasing_period_end_date()}: Returns the end date of the current leasing period.
}
}
\details{
The pre-lease season starts on September 1st and ends on July 31st of the following year.
The actual leasing period begins on August 1st.

If the current date is before September 1st, the functions will return dates for the upcoming
pre-lease season. If it's September 1st or later, they'll return dates for the pre-lease season
that has just begun.

This approach aligns with typical academic calendars, where new leases often begin around
the start of the fall semester (August 1st), and pre-leasing for the next year starts shortly after.
}
\examples{
get_pre_lease_year()
get_pre_lease_season()
get_pre_lease_season_start_date()
get_leasing_period_end_date()
}
\seealso{
Other Date Utilities: 
\code{\link{date_formatting}},
\code{\link{get_weeks_left_to_lease}()},
\code{\link{leasing_period}},
\code{\link{leasing_week}},
\code{\link{month_dates}},
\code{\link{utils_dates}},
\code{\link{weekly_period}}
}
\concept{Date Utilities}
