% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_pipeline.R
\name{prepare_pre_lease_summary_data_by_unit}
\alias{prepare_pre_lease_summary_data_by_unit}
\title{Prepare Pre-Lease Summary Data by Unit}
\usage{
prepare_pre_lease_summary_data_by_unit(summary_data)
}
\arguments{
\item{summary_data}{The processed Pre-Lease Summary Data.}
}
\value{
A tibble containing the prepared Pre-Lease Summary Data by Unit.
}
\description{
This function prepares the Pre-Lease Summary Data by Unit for further analysis and visualization. It calculates
various derived metrics and KPIs, such as occupancy rates, pre-leased percentages, year-over-year variances, and
projected availability.
}
