% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{preview_app_theme}
\alias{preview_app_theme}
\title{Preview App Theme}
\usage{
preview_app_theme(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[bslib:bs_theme_preview]{bslib::bs_theme_preview}}
  \describe{
    \item{\code{theme}}{A \code{\link[bslib:bs_theme]{bs_theme()}} object.}
    \item{\code{with_themer}}{whether or not to run the app with \code{\link[bslib:run_with_themer]{run_with_themer()}}.}
  }}
}
\value{
A preview of the GMH DataHub Shiny app theme.
}
\description{
This function previews the GMH DataHub Shiny app theme via \code{\link[bslib:bs_theme_preview]{bslib::bs_theme_preview()}}.
}
