% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_pipeline.R
\name{process_pre_lease_summary_data}
\alias{process_pre_lease_summary_data}
\title{Process Pre-Lease Summary Data}
\usage{
process_pre_lease_summary_data(summary_data, report_date = Sys.Date())
}
\arguments{
\item{summary_data}{The original Pre-Lease Summary Data.}

\item{report_date}{The report date. Default is the current system date.}
}
\value{
A tibble containing the processed Pre-Lease Summary Data.
}
\description{
This function processes the Pre-Lease Summary Data returned from the \code{pre_lease} Entrata API report summary
dataset. It performs data type conversion, column renaming, and reordering.
}
