% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{process_property_summaries}
\alias{process_property_summaries}
\title{Process all survey data and extract property summaries}
\usage{
process_property_summaries(
  all_survey_data,
  valid_prop_names = NULL,
  output_path = NULL
)
}
\arguments{
\item{all_survey_data}{The loaded survey data}

\item{valid_prop_names}{Vector of property names that aren't competitors}

\item{output_path}{Path to save the results}
}
\value{
Data frame of processed property summaries
}
\description{
Process all survey data and extract property summaries
}
