% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{process_rents_data}
\alias{process_rents_data}
\title{Extract and process rents by floorplan data}
\usage{
process_rents_data(all_survey_data, output_path = NULL)
}
\arguments{
\item{all_survey_data}{The loaded survey data}

\item{output_path}{Path to save the results}
}
\value{
Data frame of processed rents data
}
\description{
Extract and process rents by floorplan data
}
