% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{process_survey_workbooks}
\alias{process_survey_workbooks}
\title{Process multiple Excel workbooks}
\usage{
process_survey_workbooks(xl_files, xl_file_sheets, market_survey_specs)
}
\arguments{
\item{xl_files}{Vector of Excel file paths}

\item{xl_file_sheets}{Named list of sheets to process for each file}

\item{market_survey_specs}{Named list of range specifications}
}
\value{
List of all survey data
}
\description{
Process multiple Excel workbooks
}
\keyword{internal}
