% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{property_summary_inputs_tbl}
\alias{property_summary_inputs_tbl}
\title{Property Summary Inputs}
\format{
A tibble with input field definitions for property summaries.
}
\usage{
property_summary_inputs_tbl
}
\description{
Input fields for property summary data collection in market surveys.
}
\keyword{datasets}
