% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{reactable_theme}
\alias{reactable_theme}
\title{Reactable Theme}
\usage{
reactable_theme(...)
}
\arguments{
\item{...}{Named arguments to pass to the Reactable theme.}
}
\value{
A \code{\link[reactable:reactableTheme]{reactable::reactableTheme()}} leveraging the GMH Communities brand assets.
}
\description{
This function generates a custom Reactable theme for the GMH DataHub Shiny app.
}
