% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{read_sheet_data}
\alias{read_sheet_data}
\title{Read a single sheet from an Excel workbook}
\usage{
read_sheet_data(wb_path, sheet, specs)
}
\arguments{
\item{wb_path}{Path to Excel workbook}

\item{sheet}{Sheet name to read}

\item{specs}{Market survey specifications}
}
\value{
List of data frames, one for each section
}
\description{
Read a single sheet from an Excel workbook
}
\keyword{internal}
