% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  ui = app_ui,
  server = app_server,
  port = 8080,
  host = "0.0.0.0",
  on_start = NULL,
  options = app_opts(port = port, host = host),
  enable_bookmarking = NULL,
  ui_pattern = ".*",
  ...
)
}
\arguments{
\item{ui, server}{The UI and server functions.}

\item{port, host}{The port and host to run the app on.}

\item{on_start}{A function to run when the app starts.}

\item{options}{A list of options to pass to the Shiny app.}

\item{enable_bookmarking}{Enable bookmarking.}

\item{ui_pattern}{A regular expression to match the UI.}

\item{...}{Additional arguments to pass to \code{shiny::shinyApp}.}
}
\value{
Runs the shiny app.
}
\description{
Run the Shiny Application.
}
