% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_pipeline.R
\name{summarize_pre_lease_details_data_by_unit}
\alias{summarize_pre_lease_details_data_by_unit}
\title{Summarize Pre-Lease Details Data by Unit}
\usage{
summarize_pre_lease_details_data_by_unit(details_data, summary_data)
}
\arguments{
\item{details_data}{The processed Pre-Lease Details Data.}

\item{summary_data}{The prepared Pre-Lease Summary Data by Unit.}
}
\value{
A tibble containing the summarized Pre-Lease Details Data by Unit.
}
\description{
This function summarizes the Pre-Lease Details Data by Unit and merges it with the Pre-Lease Summary Data.
}
