% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survey_status.R
\name{tbl_survey_status}
\alias{tbl_survey_status}
\title{Survey Status Table}
\usage{
tbl_survey_status(data)
}
\arguments{
\item{data}{A data frame containing the following columns: \code{name}, \code{type},
\code{last_survey_date}, and \code{survey_status}.}
}
\value{
A \code{\link[reactable:reactable]{reactable::reactable()}} object.#'
}
\description{
This function renders a \code{\link[reactable:reactable]{reactable::reactable()}} with survey status and
last survey date information by property and competitor for use
in \link{mod_survey_admin}.
}
