% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gcs.R
\name{upload_entrata_report_params}
\alias{upload_entrata_report_params}
\title{Upload Entrata API Report Parameters to GCS}
\usage{
upload_entrata_report_params(report_params, report_name)
}
\arguments{
\item{report_params}{The report parameters to upload.}

\item{report_name}{The name of the report.}
}
\value{
If successful, a metadata object is returned from \code{\link[googleCloudStorageR:gcs_upload]{googleCloudStorageR::gcs_upload()}}.
}
\description{
This function uploads the Entrata API report parameters JSON to Google Cloud Storage.
}
