% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gcs.R
\name{upload_entrata_request}
\alias{upload_entrata_request}
\alias{upload_entrata_response}
\title{Upload Entrata API Request/Response JSON to GCS}
\usage{
upload_entrata_request(req, endpoint, method)

upload_entrata_response(resp, endpoint, method)
}
\arguments{
\item{req}{The request object to upload.}

\item{endpoint}{The Entrata API endpoint.}

\item{method}{The Entrata API method name.}

\item{resp}{The response object to upload.}
}
\value{
If successful, a metadata object is returned from \code{\link[googleCloudStorageR:gcs_upload]{googleCloudStorageR::gcs_upload()}}.
}
\description{
These functions upload the Entrata API request and response JSON to Google Cloud Storage.
\itemize{
\item \code{upload_entrata_request()}: Upload Entrata API request JSON to GCS.
\item \code{upload_entrata_response()}: Upload Entrata API response JSON to GCS.
}
}
