% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gcs.R
\name{upload_property_image}
\alias{upload_property_image}
\title{Upload GMH Property Image to GCS}
\usage{
upload_property_image(file, property_id, image_type = "property")
}
\arguments{
\item{file}{The file path to the image to upload.}

\item{property_id}{The property ID to associate the image with.}

\item{image_type}{The type of image to upload. Defaults to "property".}
}
\value{
If successful, a metadata object is returned from \code{\link[googleCloudStorageR:gcs_upload]{googleCloudStorageR::gcs_upload()}}.
}
\description{
This function uploads a GMH property image to Google Cloud Storage.
}
