% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{utils_checks}
\alias{utils_checks}
\alias{check_inherits}
\alias{check_chart}
\alias{check_tibble}
\alias{check_request}
\alias{check_response}
\alias{check_reactable}
\alias{check_tag}
\alias{check_taglist}
\alias{check_bslib_page}
\alias{check_percent}
\alias{check_date}
\alias{check_entrata_date}
\alias{check_entrata_daterange}
\alias{check_installed}
\title{Check Utility Functions}
\usage{
check_inherits(
  obj,
  class,
  obj_arg = rlang::caller_arg(obj),
  call = rlang::caller_env()
)

check_chart(chart, arg = rlang::caller_arg(chart), call = rlang::caller_env())

check_tibble(tbl, arg = rlang::caller_arg(tbl), call = rlang::caller_env())

check_request(req, arg = rlang::caller_arg(req), call = rlang::caller_env())

check_response(resp, arg = rlang::caller_arg(resp), call = rlang::caller_env())

check_reactable(tbl, arg = rlang::caller_arg(obj), call = rlang::caller_env())

check_tag(obj, arg = rlang::caller_arg(obj), call = rlang::caller_env())

check_taglist(obj, arg = rlang::caller_arg(obj), call = rlang::caller_env())

check_bslib_page(obj, arg = rlang::caller_arg(obj), call = rlang::caller_env())

check_percent(val, arg = rlang::caller_arg(val), call = rlang::caller_env())

check_date(date, arg = rlang::caller_arg(date), call = rlang::caller_env())

check_entrata_date(
  date,
  arg = rlang::caller_arg(date),
  call = rlang::caller_env()
)

check_entrata_daterange(
  daterange,
  arg = rlang::caller_arg(daterange),
  call = rlang::caller_env()
)

check_installed(pkg, arg = rlang::caller_arg(pkg), call = rlang::caller_env())
}
\arguments{
\item{obj, chart, tbl, req, resp, val}{The object to check the class of.}

\item{class}{The class to check against.}

\item{arg, obj_arg, call}{\link[rlang:args_error_context]{rlang::args_error_context} Arguments for error messaging.}
}
\description{
A collection of functions for checking the validity of various R objects,
primarily used for argument validation in functions.
}
\details{
Functions:
}
