% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{utils_excel_pre_lease}
\alias{utils_excel_pre_lease}
\alias{xl_read_pre_lease_global_summary_tbl}
\alias{xl_read_pre_lease_property_summary_tbls}
\alias{xl_read_pre_lease_property_details_tbls}
\alias{xl_read_pre_lease_report_parameters}
\title{Pre-Lease Excel Data}
\usage{
xl_read_pre_lease_global_summary_tbl(xl_file)

xl_read_pre_lease_property_summary_tbls(xl_file)

xl_read_pre_lease_property_details_tbls(xl_file, charge_codes = FALSE)

xl_read_pre_lease_report_parameters(xl_file)
}
\arguments{
\item{xl_file}{Path to the Excel file.}
}
\value{
\itemize{
\item \code{xl_read_pre_lease_global_summary_tbl()}: A data frame containing the global summary table.
\item \code{xl_read_pre_lease_property_summary_tbls()}: A data frame containing the merged property summary tables.
\item \code{xl_read_pre_lease_property_details_tbls()}: A data frame containing the merged property details tables.
\item \code{xl_read_pre_lease_report_parameters()}: A data frame containing the report parameters.
}
}
\description{
Set of functions to read and process GMH Pre-Lease Excel data.
\itemize{
\item \code{xl_read_pre_lease_global_summary_tbl()}: Reads the global summary table from the Excel file.
\item \code{xl_read_pre_lease_property_summary_tbls()}: Reads the property summary tables from the Excel file.
\item \code{xl_read_pre_lease_property_details_tbls()}: Reads the property details tables from the Excel file.
\item \code{xl_read_pre_lease_report_parameters()}: Reads the report parameters from the Excel file.
}
}
