% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{validate_col_names}
\alias{validate_col_names}
\title{Validate Column Names}
\usage{
validate_col_names(
  data,
  req_cols,
  optional_cols = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{The data frame to validate against.}

\item{req_cols}{Character vector of required column names.}

\item{optional_cols}{Character vector of optional column names.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Returns invisibly if the data frame has the required columns.
}
\description{
This function validates the column names of a data frame. It is used
throughout the package to ensure that the data frame has the required
columns before attempting to use or process the data.
}
\examples{
validate_col_names(mtcars, c("mpg", "cyl"))
}
