% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_entrata_config}
\alias{validate_entrata_config}
\title{Validate Entrata Configuration}
\usage{
validate_entrata_config(
  cfg,
  arg = rlang::caller_arg(cfg),
  call = rlang::caller_env()
)
}
\arguments{
\item{cfg}{A list containing the Entrata configuration values or a character
string representing the path to the configuration file.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Will throw an error if the configuration is invalid, otherwise
will invisibly return the validated \code{cfg}.
}
\description{
This function validates the Entrata configuration to ensure that all
required configuration values are present and that they are of the correct
type.

If the configuration is invalid, an error will be thrown in the caller
function's caller environment.
}
\details{
The following validations are performed on the Entrata configuration:
\enumerate{
\item The Entrata configuration must contain the following keys:
\code{username}, \code{password}, and \code{base_url}.
\item All required keys must not be empty or \code{NULL}.
\item The \code{base_url} key must be a valid \code{entrata.com} domain URL.
}

To validate the credentials provided in the configuration use the
\code{\link[=entrata_status]{entrata_status()}} function which pings the API to check the status.
}
\seealso{
\code{\link[=get_entrata_config]{get_entrata_config()}} - a function that retrieves the Entrata
configuration from a configuration file.

Other Configuration: 
\code{\link{get_entrata_config}()}

Other Validation: 
\code{\link{validate_entrata_request}()}

Other Entrata: 
\code{\link{get_entrata_config}()},
\code{\link{validate_entrata_request}()}
}
\concept{Configuration}
\concept{Entrata}
\concept{Validation}
