% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_entrata_request}
\alias{validate_entrata_request}
\alias{validate_entrata_request_endpoint}
\alias{validate_entrata_request_endpoint_method}
\alias{validate_entrata_endpoint_method_params}
\alias{validate_entrata_method_version}
\title{Validate Entrata API Request}
\usage{
validate_entrata_request(
  endpoint,
  method,
  method_params,
  method_version = NULL,
  arg_endpoint = rlang::caller_arg(endpoint),
  arg_method = rlang::caller_arg(method),
  arg_method_params = rlang::caller_arg(method_params),
  arg_method_version = rlang::caller_arg(method_version),
  call = rlang::caller_env()
)

validate_entrata_request_endpoint(
  endpoint,
  arg = rlang::caller_arg(endpoint),
  call = rlang::caller_env()
)

validate_entrata_request_endpoint_method(
  endpoint,
  method,
  arg_endpoint = rlang::caller_arg(endpoint),
  arg_method = rlang::caller_arg(method),
  call = rlang::caller_env()
)

validate_entrata_endpoint_method_params(
  endpoint,
  method,
  method_params,
  arg_endpoint = rlang::caller_arg(endpoint),
  arg_method = rlang::caller_arg(method),
  arg_method_params = rlang::caller_arg(method_params),
  call = rlang::caller_env()
)

validate_entrata_method_version(
  endpoint,
  method,
  method_version,
  arg_endpoint = rlang::caller_arg(endpoint),
  arg_method = rlang::caller_arg(method),
  arg_method_version = rlang::caller_arg(method_version),
  call = rlang::caller_env()
)
}
\arguments{
\item{endpoint}{The Entrata API endpoint to validate.}

\item{method}{The method to validate for the provided endpoint.}

\item{method_params}{A list of parameters to validate for the method.}

\item{method_version}{The version of the method to validate. Default is \code{NULL}.}

\item{arg_endpoint, arg_method, arg_method_params, arg_method_version}{Internal
arguments for capturing the caller's argument names.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}
}
\value{
All validation functions are used for their side effects and have
no return values, except for the main \code{validate_entrata_request()} function,
which will invisibly return \code{TRUE} if the request is valid.
}
\description{
This function validates an Entrata API request to ensure that the provided
endpoint, method, method parameters, and method version are valid.
}
\details{
This function is used to validate an Entrata API request before sending it
to the Entrata API. It checks that the provided endpoint is valid, that the
method is valid for the endpoint, that the method parameters are valid for
the method, and that the method version is valid if provided.
}
\seealso{
\href{https://docs.entrata.com/api/v1/documentation}{Entrata API Documentation}

Other Entrata: 
\code{\link{get_entrata_config}()},
\code{\link{validate_entrata_config}()}

Other Validation: 
\code{\link{validate_entrata_config}()}
}
\concept{Entrata}
\concept{Validation}
