% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{weekly_period}
\alias{weekly_period}
\alias{get_weekly_period}
\alias{get_weekly_period_start_date}
\title{Weekly Period}
\usage{
get_weekly_period(as_of_date = Sys.Date())

get_weekly_period_start_date(as_of_date = Sys.Date())
}
\arguments{
\item{as_of_date}{Date to use as the reference point for deriving other dates from.
Defaults to the current system date.}
}
\value{
\itemize{
\item \code{get_weekly_period()}: A list containing the start and end dates of the weekly period.
\item \code{get_weekly_period_start_date()}: The start date of the weekly period.
}
}
\description{
These functions handle calculations related to weekly periods, which are
typically used for reporting purposes. A weekly period is defined as a
period of seven days.

The following functions are defined:
\itemize{
\item \code{get_weekly_period()}: Returns the start and end dates of the weekly period.
\item \code{get_weekly_period_start_date()}: Returns the start date of the weekly period.
}

There is no need for a function to calculate the end date of the weekly period,
as it is simply the "as of" date.
}
\examples{
get_weekly_period()
get_weekly_period_start_date()

}
\seealso{
\link{leasing_week}, \link{date_formatting}

Other Date Utilities: 
\code{\link{date_formatting}},
\code{\link{get_weeks_left_to_lease}()},
\code{\link{leasing_period}},
\code{\link{leasing_week}},
\code{\link{month_dates}},
\code{\link{pre_lease_season}},
\code{\link{utils_dates}}
}
\concept{Date Utilities}
