% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{with_loader}
\alias{with_loader}
\title{With Loader}
\usage{
with_loader(ui_element, ...)
}
\arguments{
\item{ui_element}{A UI element that should be wrapped with a loader when the corresponding output is being calculated.}

\item{...}{Additional arguments to pass to \code{\link[shinycustomloader:withLoader]{shinycustomloader::withLoader()}}.}
}
\value{
The output or UI element with the loader added.
}
\description{
Add a Loader to a Shiny Output or UI Element.
}
\seealso{
\code{\link[shinycustomloader:withLoader]{shinycustomloader::withLoader()}}
}
