% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{with_tooltip}
\alias{with_tooltip}
\title{With Tooltip}
\usage{
with_tooltip(
  input,
  tooltip_text,
  placement = c("auto", "top", "right", "bottom", "left"),
  ...
)
}
\arguments{
\item{input}{The input to add the tooltip to.}

\item{tooltip_text}{The text to display in the tooltip.}

\item{placement}{The placement of the tooltip relative to its trigger.}

\item{...}{Additional arguments to pass to \code{\link[bslib:tooltip]{bslib::tooltip()}}.}
}
\value{
The input with the tooltip added.
}
\description{
Add a Tooltip to a Shiny Input.
}
\examples{
shiny::selectInput(
  inputId = "example",
  label = "Example",
  choices = c("A", "B", "C"),
  selected = "A"
) |>
  with_tooltip("This is an example tooltip.")
}
\seealso{
\code{\link[bslib:tooltip]{bslib::tooltip()}}
}
