% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{xl_get_cells}
\alias{xl_get_cells}
\title{Get Cells}
\usage{
xl_get_cells(file, sheets = NULL)
}
\arguments{
\item{file}{The path to the Excel file.}

\item{sheets}{A vector of sheet names to read. If \code{NULL}, all sheets will be read.}
}
\value{
A data frame containing cell data.
}
\description{
This function retrieves cell data from an Excel file using the \code{tidyxl} package.
}
