% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel.R
\name{xl_read_market_survey_data}
\alias{xl_read_market_survey_data}
\title{Load Market Survey Data from the Excel Workbook Template}
\usage{
xl_read_market_survey_data(
  root_xl_path,
  ignore_sheets = NULL,
  output_cache = NULL
)
}
\arguments{
\item{root_xl_path}{Root directory containing Excel files}

\item{ignore_sheets}{Vector of sheet names to ignore. Defaults to a set of common sheets if left \code{NULL}.}

\item{output_cache}{Path to save cached data (optional)}
}
\value{
List of survey data organized by file and sheet
}
\description{
This function loads market survey data from Excel files in a specified directory.
}
