% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_assets.R
\name{add_external_resources}
\alias{add_external_resources}
\title{Add External Resources}
\usage{
add_external_resources()
}
\value{
Adds external resources to the shiny app.
}
\description{
Add external resources to the shiny app.
}
