% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Shiny App UI}
\usage{
app_ui(req = NULL)
}
\arguments{
\item{req}{(internal) The request object.}
}
\value{
The UI for the Shiny application.
}
\description{
This is the main UI function for the Shiny application.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::shinyApp(ui = app_ui, server = app_server)
\dontshow{\}) # examplesIf}
}
\seealso{
Other \strong{Shiny App}: 
\code{\link{app_server}()}
}
\concept{\strong{Shiny App}}
