% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_custom_sign_in.R
\docType{data}
\name{custom_sign_in_ui}
\alias{custom_sign_in_ui}
\title{Custom Sign In UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
custom_sign_in_ui
}
\description{
the custom Sign In UI for \code{noClocksAuthR}
}
\keyword{datasets}
