% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{db_read_sql}
\alias{db_read_sql}
\title{Read and Execute SQL File}
\usage{
db_read_sql(pool, sql_file, ...)
}
\arguments{
\item{pool}{A database connection pool object.}

\item{sql_file}{Path to the SQL file.}

\item{...}{Additional arguments passed to \code{dbExecute()}.}
}
\value{
The result of the SQL execution.
}
\description{
This function reads an SQL file and executes the SQL code on the database poolection.
}
