% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_request.R
\name{entrata_request}
\alias{entrata_request}
\title{Entrata API HTTP Request}
\usage{
entrata_request(
  entrata_config = get_entrata_config(),
  endpoint = NULL,
  method_name = NULL,
  method_version = NULL,
  method_params = list(),
  request_id = NULL
)
}
\arguments{
\item{entrata_config}{Entrata API Configuration values as a list. The configuration
values should include the following keys: \code{username}, \code{password}, and \code{base_url}
at a minimum. If left \code{NULL}, will retrieve the configuration via the
\code{\link[=get_entrata_config]{get_entrata_config()}} function.}

\item{endpoint}{The Entrata API endpoint to send the request to.
This must be a valid Entrata API endpoint path. If left \code{NULL},
no endpoint will be appended to the request's URL. See
details for how to get the available Entrata API endpoints.}

\item{method_name}{The name of the Entrata API "method" (operation) to use
in the request. Not to be confused with the \code{HTTP} request method
(i.e. \code{GET}, \code{POST}, etc.), this method must be a method available for
the specified \code{endpoint} argument, making it conditional on the supplied
\code{endpoint}. If left \code{NULL}, a pre-defined default method will be used
depending on the supplied \code{endpoint}. For example, the default method
for the \verb{/properties} endpoint is \code{getProperties}. See details for how
to get the available Entrata API methods by their endpoint.}

\item{method_version}{The version of the Entrata API method to use in the
request. This is conditional on the supplied \code{method_name} argument
and must be a valid version for the given method name. If left \code{NULL},
a pre-defined default version will be used depending on the supplied
\code{method_name}. Note that all Entrata API method versions are one of
\code{"r1"}, \code{"r2"}, or \code{"r3"}.}

\item{method_params}{A named, potentially nested list of request method parameters
to pass to the Entrata API. These parameters will be added to the request
body's \code{method.params} object. Depending on the supplied \code{endpoint},
\code{method_name}, and \code{method_version}, the required parameters will vary
and some will have required, optional, or default values. If left \code{NULL},
any required method parameters without default values will throw an error.}

\item{request_id}{The unique identifier for the request.
This is used to track the request and response.  If left \code{NULL}, a
unique identifier will be generated via \code{as.integer(Sys.time())}.}
}
\value{
\code{\link[httr2:request]{httr2::request()}} object
}
\description{
Create a generic HTTP request to the Entrata API. This function initializes
and modifies a base request to the Entrata API by calling \code{\link[=entrata_req_init]{entrata_req_init()}},
\code{\link[=entrata_req_id]{entrata_req_id()}}, \code{\link[=entrata_req_endpoint]{entrata_req_endpoint()}}, and \code{\link[=entrata_req_body]{entrata_req_body()}}
as necessary to derive the final request object.

This functions primary purpose is to combine the functionality behind the
\code{\link[=entrata_req_endpoint]{entrata_req_endpoint()}} and \code{\link[=entrata_req_body]{entrata_req_body()}} functions into a single
function call allowing the end-developer to specify both the endpoint path
and its corresponding request body method details in a single function call.
}
\seealso{
\href{https://docs.entrata.com/api/v1/documentation/}{Entrata API Documentation}

\code{\link[httr2:request]{httr2::request()}}

\code{\link[=entrata_req_init]{entrata_req_init()}}, \code{\link[=entrata_req_endpoint]{entrata_req_endpoint()}}, and \code{\link[=entrata_req_body]{entrata_req_body()}}
}
