% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrata_response.R
\name{entrata_response}
\alias{entrata_response}
\title{Entrata Response}
\arguments{
\item{resp}{An \code{\link[httr2:response]{httr2::response()}} object.}
}
\value{
\itemize{
\item \code{entrata_resp_body()} Returns a parsed list representing the API response's JSON.
}
}
\description{
Functions for working with API responses from the Entrata API.
}
\seealso{
\code{\link[=entrata_request]{entrata_request()}}

Other Entrata: 
\code{\link{entrata_config}},
\code{\link{entrata_reports}()}
}
\concept{Entrata}
