% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{icon_text}
\alias{icon_text}
\alias{text_icon}
\title{Icon Text}
\usage{
icon_text(icon, text, .function = shiny::icon)

text_icon(text, icon, .function = shiny::icon)
}
\arguments{
\item{icon}{Character string for the icon name.}

\item{text}{Character string for the text.}

\item{.function}{Function to create the icon. Defaults to \code{\link[shiny:icon]{shiny::icon()}}.}
}
\value{
HTML element with the icon and text.
}
\description{
Create an HTML element with an icon and text.
\itemize{
\item \code{icon_text()}: Icon followed by text.
\item \code{text_icon()}: Text followed by icon.
}
}
\examples{
icon_text("home", "Home")
text_icon("Home", "home")
}
