% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{mod_home}
\alias{mod_home}
\alias{mod_home_ui}
\alias{mod_home_server}
\alias{mod_home_demo}
\title{Home Shiny Module}
\usage{
mod_home_ui(id)

mod_home_server(id, pool = NULL, global_filters = NULL)

mod_home_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_home_ui()}: UI output
\item \code{mod_home_server()}: List of reactive expressions.
\item \code{mod_home_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Home page.
\itemize{
\item \code{mod_home_ui()}: The front-end UI for the module.
\item \code{mod_home_server()}: The server logic for the module.
\item \code{mod_home_demo()}: Function to demo the module in an isolated shiny app context.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_home_demo()
\dontshow{\}) # examplesIf}
}
