% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pre_lease.R
\name{mod_pre_lease}
\alias{mod_pre_lease}
\alias{mod_pre_lease_ui}
\alias{mod_pre_lease_server}
\alias{mod_pre_lease_demo}
\title{Pre-Lease Shiny Module}
\usage{
mod_pre_lease_ui(id)

mod_pre_lease_server(id, pool = NULL)

mod_pre_lease_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}

\item{selected_property}{Selected property ID.}

\item{report_date}{Report date.}
}
\value{
\itemize{
\item \code{mod_pre_lease_ui()}: UI HTML Output.
\item \code{mod_pre_lease_server()}: List of reactive values.
\item \code{mod_pre_lease_demo()}: \code{NULL}, used for the side-effect of a demo app.
}
}
\description{
A Shiny Module for the GMH Communities Pre-Lease Summary & Details reports
data visualization and review.

The following functions are implemented:
\itemize{
\item \code{mod_pre_lease_ui()}: User interface (UI) definition
\item \code{mod_pre_lease_server()}: Server logic
\item \code{mod_pre_lease_demo()}: Demo application for the module
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_pre_lease_demo()
\dontshow{\}) # examplesIf}
}
