% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_properties.R
\name{mod_properties}
\alias{mod_properties}
\alias{mod_properties_ui}
\alias{mod_properties_server}
\alias{mod_properties_demo}
\title{Properties Shiny Module}
\usage{
mod_properties_ui(id)

mod_properties_server(id, pool = NULL, global_filters = NULL)

mod_properties_demo()
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}
}
\value{
\itemize{
\item \code{mod_properties_ui()}: UI output
\item \code{mod_properties_server()}: List of reactive expressions.
\item \code{mod_properties_demo()}: \code{NULL}, runs the demo application.
}
}
\description{
A Shiny Module for the GMH Data Hub's Properties page.

Includes the following functions:
\itemize{
\item \code{mod_properties_ui()}: User Interface (UI) for the module.
\item \code{mod_properties_server()}: Server logic for the module.
\item \code{mod_properties_demo()}: Demo application for the module.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_properties_demo()
\dontshow{\}) # examplesIf}
}
