% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_survey_notes.R
\name{mod_survey_notes}
\alias{mod_survey_notes}
\alias{mod_survey_notes_ui}
\alias{mod_survey_notes_server}
\alias{create_note_card}
\title{Survey Notes Section Module}
\usage{
mod_survey_notes_ui(id)

mod_survey_notes_server(
  id,
  pool = NULL,
  survey_data = NULL,
  selected_filters = NULL,
  db_trigger_func = NULL,
  edit_survey_section = NULL
)

create_note_card(note, ns)
}
\arguments{
\item{id}{Module's namespace ID.}

\item{pool}{Database connection pool.}

\item{note}{A single note object.}

\item{ns}{Namespace function.}

\item{global_filters}{Reactive expression for global filters.}
}
\value{
\itemize{
\item \code{mod_survey_notes_ui()}: UI output.
\item \code{mod_survey_notes_server()}: List of reactive expressions.
\item \code{mod_survey_notes_demo()}: \code{NULL}, runs a demo application.
\item \code{create_note_card()}: Utility function to create note cards.
}
}
\description{
A Shiny Module for the GMH Data Hub's Market Survey "Notes" Section.

Includes the following functions:
\itemize{
\item \code{mod_survey_notes_ui()}: User Interface (UI) for the module.
\item \code{mod_survey_notes_server()}: Server logic for the module.
\item \code{mod_survey_notes_demo()}: Demo application for the module.
}

plus an additional UI utility function to create note cards:
\itemize{
\item \code{create_note_card()}: Create a note card for display.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_survey_notes_demo()
\dontshow{\}) # examplesIf}
}
