% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg.R, R/utils_pkgdown.R
\name{pkg_docs}
\alias{pkg_docs}
\title{Package Documentation}
\usage{
pkg_docs()

pkg_docs()
}
\value{
Opens the package documentation in the default web browser.
}
\description{
Open the package documentation in the default web browser.

This function launches the \code{pkgdown} documentation for the \code{gmhdatahub}
package.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pkg_docs()
\dontshow{\}) # examplesIf}
}
