% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkg.R
\name{pkg_sys}
\alias{pkg_sys}
\title{Package System File}
\usage{
pkg_sys(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[base:system.file]{base::system.file()}}}
}
\value{
The path to the system file.
}
\description{
Package System File
}
\examples{
pkg_sys("extdata", "data.csv")
}
